% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_characteristics.R
\name{add_char_demo}
\alias{add_char_demo}
\title{Add characteristic by matching on demographics}
\usage{
add_char_demo(synth_pop_fn, output_path, pop_type, args)
}
\arguments{
\item{synth_pop_fn}{full name of the synth pop}

\item{output_path}{the path to the folder of where the new output is to go}

\item{pop_type}{"b" for both; "household", or "people" populations}

\item{args}{list of additional arguments including the suffix}
}
\value{
logical, writes out new synth pop)
}
\description{
Add characteristic by matching on demographics
}
