% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_characteristics.R
\name{add_characteristic}
\alias{add_characteristic}
\title{Add a characteristic to an existing population}
\usage{
add_characteristic(synth_pop_path, args, pop_type = "people",
  method = "demo_matching", doPar = FALSE)
}
\arguments{
\item{synth_pop_path}{path to folder where the synth pop(s) is/are}

\item{args}{list of additional arguments}

\item{pop_type}{either "household" or "people"  population or "both" for household characteristics joined to people}

\item{method}{"demo_matching"}

\item{doPar}{logical, default is FALSE}
}
\value{
a written synth pop with added characteristics
}
\description{
Add a characteristic to an existing population
}

