% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{align_pums}
\alias{align_pums}
\title{Match the pums variables with marginal totals}
\usage{
align_pums(pums, marginals)
}
\arguments{
\item{pums}{dataframe subsetted to only the marginal vars}

\item{marginals}{list containing all of the marginal totals}
}
\value{
pums dataframe with the marginal columns binded on
}
\description{
Match the pums variables with marginal totals
}

