% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{allocate_count}
\alias{allocate_count}
\title{Re-allocate excess counts to other locations}
\usage{
allocate_count(counts, count_id)
}
\arguments{
\item{counts}{numeric vector of current counts}

\item{count_id}{numeric index indicating which 
count is excess}
}
\value{
new_counts a new numeric count vector with the
}
\description{
Re-allocate excess counts to other locations
}
