% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{append_pops}
\alias{append_pops}
\title{Append populations within a directory into two separate files:  household and people populations.}
\usage{
append_pops(input_dir, output_dir = input_dir, remove_pops = FALSE)
}
\arguments{
\item{input_dir}{path to the directory (bottom level) of populations to append}

\item{output_dir}{path to output directory}

\item{remove_pops}{TRUE if we should we remove the smaller populations}
}
\value{
TRUE
}
\description{
Append populations within a directory into two separate files:  household and people populations.
}

