% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{assessSynecosUS}
\alias{assessSynecosUS}
\title{Assess the environment use of our synthetic ecosystems}
\usage{
assessSynecosUS(syneco_list, type = "both")
}
\arguments{
\item{syneco_list}{output from makeEnvironmentsDFs}

\item{type}{"wpl" "sch" or "both" currently}
}
\description{
Assess the environment use of our synthetic ecosystems
}
