% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{assign_place}
\alias{assign_place}
\title{Assign a place to a person}
\usage{
assign_place(people, data_list)
}
\arguments{
\item{people}{data frame of synthetic people produced by SPEW}

\item{data_list}{of the data and identifying name}
}
\value{
column corresponding to the people of the place assignment
}
\description{
Assign a place to a person
}
