% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{assign_place_coords}
\alias{assign_place_coords}
\title{Assign a place with long/lat coords to a synthetic population}
\usage{
assign_place_coords(pop, places, place_name = "place", method = "uniform",
  dist_fxn = haversine_dist)
}
\arguments{
\item{pop}{data frame with "longitude" and "latitude" columns}

\item{places}{data frame with "longitude" and "latitude" and an "ID" column, perhaps a "capacity" column}

\item{place_name}{string that will become the column name of the place}

\item{method}{c("uniform", "capacity") The method on how we assign places}

\item{dist_fxn}{(haversine_dist) or a function with args x1, y1, x2, y2}
}
\value{
pop data frame with column of place_name with the place_ids
}
\description{
Assign a place with long/lat coords to a synthetic population
}

