% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\name{assign_schools}
\alias{assign_schools}
\title{Assign schools to a synthetic population.}
\usage{
assign_schools(people, schools, weightSchools = weight_dists,
  distFun = haversine)
}
\arguments{
\item{people}{data frame of synthetic people}

\item{schools}{list with public and private school data frames}
}
\value{
a column of school assignments by school ID
}
\description{
Assign schools to a synthetic population.
}
\references{
See PUMS CODEBOOK:  http://www2.census.gov/programs-surveys/acs/tech_docs/pums/data_dict/PUMS_Data_Dictionary_2010-2014.pdf
}
