% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\name{assign_schools_inner}
\alias{assign_schools_inner}
\title{Function which assigns schools}
\usage{
assign_schools_inner(df, schools, weightSchools, distFun)
}
\arguments{
\item{df}{subset of people split so all age, grade, SCH, and county should be the same in the df}

\item{schools}{list of schools, one data frame of private and one of public}

\item{weightSchools}{Function to weight the schools}

\item{distFun}{Function to compute distance between schools and agents}
}
\value{
column of school ID assignments
}
\description{
Function which assigns schools
}
