% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{assign_weights}
\alias{assign_weights}
\title{Assign weights for ipf-based sampling}
\usage{
assign_weights(pums, table, alpha, k, marginals)
}
\arguments{
\item{pums}{dataframe with marginal columns}

\item{table}{dataframe with marginal combinations and frequencies}

\item{alpha}{number between 0 and 1, weight of categorical variables}

\item{k}{number between 0 and 1, weight of orginal variables}

\item{marginals}{list with marginal data}
}
\value{
weight matrix, rows = rows in pums, columns = rows in table
}
\description{
Assign weights for ipf-based sampling
}

