% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workplaces.R
\name{assign_workplaces}
\alias{assign_workplaces}
\title{Assign an ESRI workplace to synthetic population}
\usage{
assign_workplaces(people, workplaces)
}
\arguments{
\item{people}{data frame of synthetic people produced by SPEW}

\item{workplaces}{data frame of workplaces from ESRI}
}
\value{
column of workplace IDs for synthetic people
}
\description{
We look at the ESR variable from the synthetic people, which is the employment status recode.  If values 1, 2, 4, and 5 are considered employed.  We then randomly assign the person inside the county, with weights based on numbers of workers at the place.  We return the ESRI ID of the workplace  Note that in this way, workers do not cross borders.
}

