% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workplaces.R
\name{assign_workplaces_inner}
\alias{assign_workplaces_inner}
\title{Function which assigns workplaces}
\usage{
assign_workplaces_inner(df, workplaces)
}
\arguments{
\item{df}{subset of people with emp either 0 or 1 and the county number}

\item{workplaces}{dataframe or ESRI schools}
}
\value{
ID of ESRI workplace or NA
}
\description{
Function which assigns workplaces
}

