% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{calc_dists}
\alias{calc_dists}
\title{Calculate distance b/w cont table row and pums}
\usage{
calc_dists(pums, table_row, alpha, k, marginals)
}
\arguments{
\item{pums}{dataframe with marginal columns}

\item{table_row}{row of table to calculate distance}

\item{alpha}{number between 0 and 1, weight of categorical variables}

\item{k}{number between 0 and 1, weight of orginal variables}

\item{marginals}{list with marginal data}
}
\description{
Calculate distance b/w cont table row and pums
}
