% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{call_spew}
\alias{call_spew}
\title{Wrapper function to read, format, and generare SPEW synthetic ecosystem}
\usage{
call_spew(input_dir, folders = NULL, data_group, output_dir,
  parallel = TRUE, sampling_method = "uniform",
  locations_method = "uniform", convert_count = FALSE,
  vars = list(household = NA, person = NA))
}
\arguments{
\item{input_dir}{character specifying input data directory}

\item{folders}{list specifying sub-directories for each data-source}

\item{output_dir}{character indication where to write ecosystems}

\item{parallel}{logical whether or not to run in parallel}

\item{sampling_method}{character of sampling}

\item{locations_method}{character vector indicating the type of location 
sampling to use, defaults to "uniform", can also be "roads".}

\item{convert_count}{logical meant to indicate if we are going to convert 
population totals to househole counts}

\item{vars}{list with two components: household and person. This specifies 
which variables to include in the corresponding PUMS data-set}

\item{data-group}{character, either "US", "ipums", or "none"}

\item{make_plots}{logical indicating whether we should make maps of 
the synthetic households.}
}
\value{
logical indicating whether or not this run of spew ended successfully
}
\description{
Wrapper function to read, format, and generare SPEW synthetic ecosystem
}

