% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{cap_default}
\alias{cap_default}
\title{How to weight the capacities of of school.}
\usage{
cap_default(cap)
}
\arguments{
\item{cap}{a single number}
}
\value{
the weighted capacity, a single number.
}
\description{
How to weight the capacities of of school.
}
\details{
this default version returns the ceiling(capacity / 10).
}
