% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{ccount}
\alias{ccount}
\title{Adjust number of households}
\usage{
ccount(convert_count = FALSE, persons, np, nh, n_house)
}
\arguments{
\item{convert_count}{logical, TRUE means make the switch}

\item{persons}{vector of number of persons in each household}

\item{np}{rows of person pums}

\item{nh}{rows of household pums}

\item{n_house}{original count of number of households}
}
\value{
updated n_house
}
\description{
Only changes things when population total represents the number 
of people (instead of households)
}
