% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{check_path}
\alias{check_path}
\title{Check the path to output to run diags}
\usage{
check_path(output_dir)
}
\arguments{
\item{output_dir}{directory to TOP level region}
}
\value{
logical
}
\description{
Check the path to output to run diags
}
