% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_place_ids}
\alias{check_place_ids}
\title{Check the Place ID's match}
\usage{
check_place_ids(id1, id2)
}
\arguments{
\item{id1}{character vector of place_ids}

\item{id2}{character vector of place_ids}
}
\value{
Either "Place ids match!" or an error
detailing why
}
\description{
Check the Place ID's match
}
