% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_pop_table}
\alias{check_pop_table}
\title{Check the pop_table has all the necessary components}
\usage{
check_pop_table(pop_table)
}
\arguments{
\item{pop_table}{}
}
\value{
Either a character "Pop table is ready!", or an 
error message detailing what went wrong
}
\description{
Check the pop_table has all the necessary components
}
