% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_puma_ids}
\alias{check_puma_ids}
\title{Check the puma id's match}
\usage{
check_puma_ids(pop_table_ids, pums_ids)
}
\arguments{
\item{pop_table_ids}{character vector of ids}

\item{pums_ids}{character vector of ids}
}
\value{
Either "Puma ids match!" or an error
detailing why
}
\description{
Check the puma id's match
}
