% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check.R
\name{check_pums}
\alias{check_pums}
\title{Check that the pums has all the required components}
\usage{
check_pums(pums)
}
\arguments{
\item{pums}{list with the household and person level
pums}
}
\value{
Either a character "Pums is ready!" or an error message
detailing what went wrong
}
\description{
Check that the pums has all the required components
}

