% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check.R
\name{check_shapefile}
\alias{check_shapefile}
\title{Check the shapefile has the necessary components}
\usage{
check_shapefile(shapefile)
}
\arguments{
\item{shapefile}{}
}
\value{
Either a character"Shapefile is ready!" or an error message
detailing what went wrong
}
\description{
Check the shapefile has the necessary components
}

