% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{check_var_names}
\alias{check_var_names}
\title{Check to see if variable names are in SPEW outputs}
\usage{
check_var_names(header_h, header_p, vars_to_sum_h, vars_to_sum_p,
  vars_to_sum_env = NULL)
}
\arguments{
\item{header_h}{character vector of household variables in household SPEW output}

\item{header_p}{character vector of people variables in household SPEW output}

\item{vars_to_sum_h}{character vector of household variables we want to summarize}

\item{vars_to_sum_p}{character vector of people variables we want to summarize}

\item{vars_to_sum_env}{character vector of variables from the person data frame which correspond to environment assignments.  Default is NULL}
}
\value{
logical
}
\description{
Check to see if variable names are in SPEW outputs
}
