% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{checkDF}
\alias{checkDF}
\title{Check if df is in the right format}
\usage{
checkDF(df, type = "coords")
}
\arguments{
\item{df}{data frame}

\item{type}{("coords").  For type "coords", we check to make sure "longitude" and "latitude" are column names}
}
\value{
logical
}
\description{
Check if df is in the right format
}

