% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.R
\name{clean_names}
\alias{clean_names}
\title{Remove whitespace, capitals, and non ASCII}
\usage{
clean_names(names)
}
\arguments{
\item{names}{character vector of names to clean}
}
\value{
names character vector of all lowercase,
non-capital and ASCII names
}
\description{
Remove whitespace, capitals, and non ASCII
}

