% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{combine_counts}
\alias{combine_counts}
\title{Combine two rows of a pop_table into one}
\usage{
combine_counts(pop_table, place1, place2)
}
\arguments{
\item{pop_table}{dataframe to update}

\item{place1}{character indicating the name of the 
place to add counts to}

\item{place2}{character indicating the name of the 
place to remove}
}
\value{
pop_table a data-frame with the place2 counts 
added to place1, and place2 removed from the pop_table
}
\description{
Combine two rows of a pop_table into one
}

