% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{combine_many_counts}
\alias{combine_many_counts}
\title{Combine multiple rows of a pop_table into one and rename the aggregated row}
\usage{
combine_many_counts(pop_table, places, new_name)
}
\arguments{
\item{pop_table}{dataframe to update}

\item{places}{character vector of places to combine together}
}
\value{
pop_table a data-frame with the places' counts combined together, and renamed by new_name
}
\description{
Combine multiple rows of a pop_table into one and rename the aggregated row
}

