% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-tables.R
\name{combinePUMS}
\alias{combinePUMS}
\title{Combine the household and people PUMS for IPF}
\usage{
combinePUMS(pums_p, pums_h, person_vars = c("AGEP", "RAC1P"),
  hh_vars = c("HINCP", "NP"))
}
\arguments{
\item{pums_p}{person pums from US ACS}

\item{pums_h}{household pums from US ACS (same year as person PUMS)}

\item{person_vars}{currently in c("AGEP", "RAC1P")}

\item{hh_vars}{currently in c("HINCP", "NP")}
}
\value{
a shortened and combined set of PUMS
}
\description{
Combine the household and people PUMS for IPF
}
\references{
For details about the PUMS variables, 
see http://www2.census.gov/programs-surveys/acs/tech_docs/pums/data_dict/PUMS_Data_Dictionary_2010-2014.pdf.
}
