% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{compareCapacities}
\alias{compareCapacities}
\title{Compare capacities of generated schools and actual schools}
\usage{
compareCapacities(agents, schools)
}
\arguments{
\item{agents}{synthetic people with school_id var}

\item{list}{of public schools df and private schools df with Students variable with capacity}
}
\value{
percent filled school (possibly over 1)
}
\description{
Compare capacities of generated schools and actual schools
}

