% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{deg2rad}
\alias{deg2rad}
\title{Converts degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{degree}{(lat/long coordinate)}
}
\value{
degree in radians
}
\description{
Converts degrees to radians
}

