% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-docs.R
\docType{data}
\name{delaware}
\alias{delaware}
\title{Input data for Keny County, Delaware}
\format{An object of class \code{list} of length 8.}
\usage{
delaware
}
\description{
Input data for Keny County, Delaware
}
\keyword{datasets}
