% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_characteristics.R
\name{demo_sample}
\alias{demo_sample}
\title{Sample extra characteristics from char pums and add them to the pop df}
\usage{
demo_sample(pop_df, char_pums, var_names, args = NULL)
}
\arguments{
\item{pop_df}{a df that has been subsetted}

\item{char_pums}{data frame of PUMS characteristics to add}

\item{var_names}{variables we want to match on}

\item{args}{list of additional arguments}
}
\description{
Sample extra characteristics from char pums and add them to the pop df
}

