% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{euclidean_dist}
\alias{euclidean_dist}
\title{Get the euclidean distance between two points (x1, y1) and (x2, y2)}
\usage{
euclidean_dist(x1, y1, x2, y2)
}
\arguments{
\item{x1}{longitude of object 1 (vector)}

\item{y1}{latitude of object 1 (vector)}

\item{x2}{longitude of object 2 (vector)}

\item{y2}{latitude of object 2 (vector)}
}
\value{
numeric
}
\description{
Get the euclidean distance between two points (x1, y1) and (x2, y2)
}

