% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{extract_st_co_tr}
\alias{extract_st_co_tr}
\title{Extract the state, county, and tract ID from a string}
\usage{
extract_st_co_tr(files)
}
\arguments{
\item{files}{character vector of the files}
}
\value{
a data frame with the state, county, and tract IDs, along with the aggregated ID
}
\description{
Extract the state, county, and tract ID from a string
}
