% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{extractHeader}
\alias{extractHeader}
\title{Extract header of a file}
\usage{
extractHeader(path)
}
\arguments{
\item{path}{a path to the csv file}
}
\value{
vector of column names, the header
}
\description{
Extract header of a file
}

