% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{extractStCoTr}
\alias{extractStCoTr}
\title{Extract the StCoTr number and put it in a df}
\usage{
extractStCoTr(paths_df)
}
\arguments{
\item{paths_df}{data frame last column has stco tr number in it}
}
\value{
data frame
}
\description{
Extract the StCoTr number and put it in a df
}
