% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{extrapolate_probs_to_pums_joint}
\alias{extrapolate_probs_to_pums_joint}
\title{Take unique probabilites for table and spread them to rest of PUMS}
\usage{
extrapolate_probs_to_pums_joint(p, n, pums, var_names, tab)
}
\arguments{
\item{p}{probabilities from solve.QP of length(unique(pums[, var_name]))}

\item{n}{matrix with unique categories  p and n correspond to one another}

\item{pums}{PUMS data frame}

\item{tab}{data frame with categories and frequency}

\item{var_name}{varaible name we are matching on}
}
\value{
probabilities for whole data frame
}
\description{
Take unique probabilites for table and spread them to rest of PUMS
}

