% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-tables.R
\name{fetchAndWriteACS}
\alias{fetchAndWriteACS}
\title{fetch and write ACS tables from acs package}
\usage{
fetchAndWriteACS(geo_list, acs_fetch_list, outPath = ".", tabID = NULL)
}
\arguments{
\item{geo_list}{list with geographies for geo.make function}

\item{acs_fetch_list}{list with acs fetch data}

\item{outPath}{filepath}

\item{tabID}{way to ID table if NULL, use table.number in acs_fetch_list}
}
\value{
df of acs and also a saved csv
}
\description{
fetch and write ACS tables from acs package
}

