% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{fill_cont_table}
\alias{fill_cont_table}
\title{Fill marginal contingency with ipf}
\usage{
fill_cont_table(pums, marginals, place_id, n_house)
}
\arguments{
\item{pums}{dataframe of aligned pums data}

\item{marginals}{list with marginal totals}

\item{place_id}{ID saying which column to extract 
the marginals from}

\item{n_house}{number of households to sample}
}
\value{
table a data-frame containing all of the 
marginal combinations and their frequencies
}
\description{
Fill marginal contingency with ipf
}
