% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fipsToPlaceName}
\alias{fipsToPlaceName}
\title{Translate FIPS number to place name}
\usage{
fipsToPlaceName(fips, level, df = us)
}
\arguments{
\item{fips}{string id
length 2 for state
length 5 for county}

\item{level}{("state", "county")}
}
\description{
Translate FIPS number to place name
}
