% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{force_counts}
\alias{force_counts}
\title{Force counts to shapefiles despite bad matches from shapefiles and pop_table names}
\usage{
force_counts(shapefile_names, pop_table, split_type = "uniform")
}
\arguments{
\item{shapefile_names}{names of the regions that will be allocated counts}

\item{pop_table}{data frame to update}

\item{split_type}{in c("uniform")}
}
\value{
new_table data frame with counts that match the shapefile, total population should match the sum of the old pop_table counts
}
\description{
Force counts to shapefiles despite bad matches from shapefiles and pop_table names
}

