% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{force_pop}
\alias{force_pop}
\title{Brute force a shapefile into a 1-1 pop table}
\usage{
force_pop(pop_table, shapefile_names)
}
\arguments{
\item{shapefile_names}{names exactly corresponding to shapefile_names}

\item{total}{total number of households in the coutry}
}
\value{
data frame that matches pop_table and shapefile names exactly, splitting the population evenly
}
\description{
Take the total population and divide it by the number of regions
}

