% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-tables.R
\name{formatID}
\alias{formatID}
\title{Format the ids of a us state/county/tract}
\usage{
formatID(geo, type)
}
\arguments{
\item{geo}{IDs}

\item{type}{"state" "county" or "tract"}
}
\value{
formatted state/county/tract ID
}
\description{
Format the ids of a us state/county/tract
}

