% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-tables.R
\name{formatPUMS_hhhAge}
\alias{formatPUMS_hhhAge}
\title{Sync the PUMS AGEP with the HHH Age var}
\usage{
formatPUMS_hhhAge(pums)
}
\arguments{
\item{pums}{combined pums}
}
\value{
pums with AGEP
}
\description{
Sync the PUMS AGEP with the HHH Age var
}
