% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-tables.R
\name{formatPUMS_hhhRace}
\alias{formatPUMS_hhhRace}
\title{Sync the PUMS AGEP with the HHH Race var}
\usage{
formatPUMS_hhhRace(pums)
}
\arguments{
\item{pums}{combined pums}
}
\value{
pums with RAC1P
}
\description{
Sync the PUMS AGEP with the HHH Race var
}

