% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-tables.R
\name{formatPUMS_hhInc}
\alias{formatPUMS_hhInc}
\title{Sync the PUMS HINCP with the HHH Race var}
\usage{
formatPUMS_hhInc(pums)
}
\arguments{
\item{pums}{combined pums}
}
\value{
pums with HINCP
}
\description{
Sync the PUMS HINCP with the HHH Race var
}
