% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-tables.R
\name{formatPUMS_hhSize}
\alias{formatPUMS_hhSize}
\title{Sync the PUMS NP with the HH Size var}
\usage{
formatPUMS_hhSize(pums)
}
\arguments{
\item{pums}{combined pums}
}
\value{
pums with NP
}
\description{
Sync the PUMS NP with the HH Size var
}
