% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-tables.R
\name{formatTab_hhSize}
\alias{formatTab_hhSize}
\title{Format tables for use in IPF}
\usage{
formatTab_hhSize(tab_obj)
}
\arguments{
\item{tab_obj}{from getting data from acs}
}
\value{
the formatted table
}
\description{
Format tables for use in IPF
}

