% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{generate_spew}
\alias{generate_spew}
\title{Generate synthetic ecosysystems}
\usage{
generate_spew(input_dir, folders, data_group, output_dir, parallel = TRUE,
  sampling_method = "uniform", locations_method = "uniform",
  convert_count = FALSE, vars = list(household = NA, person = NA))
}
\arguments{
\item{input_dir}{character vector specifying the directory containing all of the 
input data}

\item{folders}{list which specifies the sub-directory for each piece of data}

\item{output_dir}{character vector indicating the directory to write 
the synthetic populations}

\item{parallel}{logical indicating whether or not the make_data function 
should be run in parallel}

\item{sampling_method}{character vector indicating the type of sampling 
method to use, defaults to "uniform"}

\item{locations_method}{character vector indicating the type of location 
sampling to use, defaults to "uniform", can also be "roads".}

\item{convert_count}{logical meant to indicate if we are going to convert 
population totals to househole counts}

\item{vars}{list with two components: household and person. This specifies 
which variables to include in the corresponding PUMS data-set}

\item{data-group}{character vector, either "US", "ipums", or "none".}

\item{make_plots}{logical indicating whether we should make maps of 
the synthetic households.}
}
\value{
logical indicating whether or not this run of spew ended successfully
}
\description{
Generate synthetic ecosysystems
}

