% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{get_base_map}
\alias{get_base_map}
\title{Get the base map for plotting}
\usage{
get_base_map(coords_df, get_world_map = TRUE, f = 0.1)
}
\arguments{
\item{coords_df}{data frame with "longitude", "latitude", and "region_id" columns}

\item{get_world_map}{logical.  Get an underlying map from ggmap.  Default is TRUE}

\item{f}{border size of map}
}
\value{
a ggplotting object
}
\description{
Get the base map for plotting
}
