% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{get_centers}
\alias{get_centers}
\title{Get the center longitude and latitude for each region}
\usage{
get_centers(coords_df)
}
\arguments{
\item{coords_df}{data frame with "longitude", "latitude", and "region_id" columns}
}
\value{
data frame with mean lon/lat and region
}
\description{
Get the center longitude and latitude for each region
}
