% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{get_coords_scaled}
\alias{get_coords_scaled}
\title{Getting a plotting data frame}
\usage{
get_coords_scaled(sum_list, samp_size, coords, pop_totals)
}
\arguments{
\item{sum_list}{output for people from summarize_spew}

\item{samp_size}{how many records to keep for plotting}

\item{coords}{logical}

\item{pop_totals}{number of people to sample from each region}
}
\value{
data frame for plotting, longitude and latitude coordinates along with region
}
\description{
Getting a plotting data frame
}
