% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{get_dfs}
\alias{get_dfs}
\title{Get the dataframes from SPEW summary output}
\usage{
get_dfs(sum_list, vars_to_sum, has_marg = FALSE)
}
\arguments{
\item{sum_list}{from summarize_spew}

\item{vars_to_sum}{variables to make data frames of}

\item{has_marg}{logical.  Does the synthetic ecosystem have marginals to look at?  Default is FALSE}
}
\value{
list of data frame for each variable
}
\description{
Get the dataframes from SPEW summary output
}
\details{
the output is very gg-verse friendly
}
