% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{get_dist_mat}
\alias{get_dist_mat}
\title{Get the distance matrix}
\usage{
get_dist_mat(pop, places, dist_fxn = haversine_dist)
}
\arguments{
\item{pop}{data frame of the population with m rows}

\item{places}{data frame of places to assign with n rows}

\item{dist_fxn}{currently "haversine" with args for x1, y1, x2, y2, returning a scalar value between 0 and 1}
}
\value{
m x n matrix with scaled distance between 0 and 1.  Eg.  Entry ij means that the scaled distance between row i from pop and row j from places is entry ij.
}
\description{
Get the distance matrix between two data frames that have "longitude" and "latitude" columns
}

