% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\name{get_dists}
\alias{get_dists}
\title{Get the distances between the schools and the people.}
\usage{
get_dists(df, schools, dist)
}
\arguments{
\item{df}{data frame of people}

\item{schools}{dataframe of subsetted schools}

\item{dist}{"haversine"}
}
\value{
distance between two points on the globe
}
\description{
Get the distances between the schools and the people.
}
