% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{get_envs}
\alias{get_envs}
\title{Gather the unique assignments for the region}
\usage{
get_envs(p_sum_list, env_vars = NULL)
}
\arguments{
\item{p_sum_list}{output from summarize_spew}

\item{env_vars}{the environment variables we want to summarize.  Default is NULL}
}
\value{
a list of unique assignments for each of the env_variables
}
\description{
Gather the unique assignments for the region
}
