% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{get_header}
\alias{get_header}
\title{Extract the header from a population}
\usage{
get_header(output_dir, type = "household")
}
\arguments{
\item{output_dir}{directory to TOP level region}

\item{type}{either "household" or "people"}
}
\value{
a character vector of the header
}
\description{
Extract the header from a population
}
