% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.R
\name{get_level}
\alias{get_level}
\title{Obtain the correct level for ipums data}
\usage{
get_level(shapefile_names, pop_table)
}
\arguments{
\item{shapefile_names}{character vector of the names
which come from the given shapefile}

\item{pop_table}{data frame in which we want to obtain the
levels}
}
\value{
level character of the level in which we will
use
}
\description{
Obtain the correct level for ipums data
}

