% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{get_pop_totals}
\alias{get_pop_totals}
\title{Get the population totals from a summarized SPEW region}
\usage{
get_pop_totals(hh_sum_list, p_sum_list)
}
\arguments{
\item{hh_sum_list}{output from summarize_spew for households}

\item{p_sum_list}{output from summarize_spew for people}
}
\value{
list of a data frame of population totals, one record for each region, number of houses, and number of people
}
\description{
Get the population totals from a summarized SPEW region
}
