% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfiles.R
\name{get_rows}
\alias{get_rows}
\title{Extract rows with a certain character}
\usage{
get_rows(dat, char)
}
\arguments{
\item{dat}{character vector}

\item{char}{character which specifies the
sequence we are looking for}
}
\value{
output which is the subsetted rows
of the initial character vector (logfile)
}
\description{
Extract rows with a certain character
}
