% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_shapefile_indices}
\alias{get_shapefile_indices}
\title{Obtain the shapefile indices corresponding to the pop table}
\usage{
get_shapefile_indices(shapefile_names, count_names)
}
\arguments{
\item{shapefile_names}{character vector with the shapefile names}

\item{count_names}{character vector with the count names}
}
\value{
numeric vector indicating the appropriate indices for 
shapefiles which correspond to the count_names
}
\description{
Obtain the shapefile indices corresponding to the pop table
}
