% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew_in_console.R
\name{get_supp_data}
\alias{get_supp_data}
\title{Format the supplementary data so it is compatible with SPEW}
\usage{
get_supp_data(supplementary_data)
}
\arguments{
\item{supplementary_data}{list of supplementary data}
}
\value{
list of supplementary data that is compatible with SPEW
}
\description{
Format the supplementary data so it is compatible with SPEW
}
