% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{get_targets}
\alias{get_targets}
\title{Obtain the target marginals for IPF}
\usage{
get_targets(marg_cols, marginals, place_id, n_house)
}
\arguments{
\item{marg_cols}{dataframe of marginal columns}

\item{marginals}{list of marginal data}

\item{place_id}{id corresponding to the rows of marginals}

\item{n_house}{number of households in this sample}
}
\value{
list with target_data and target list for this particular
place_id
}
\description{
Obtain the target marginals for IPF
}

