% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfiles.R
\name{get_total_time}
\alias{get_total_time}
\title{Extract the total run-time from a SPEW log-file}
\usage{
get_total_time(logfile)
}
\arguments{
\item{SPEW}{log-file}
}
\description{
Extract the total run-time from a SPEW log-file
}

