% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_var_names}
\alias{get_var_names}
\title{Get the proper names to subset}
\usage{
get_var_names(type, summary_vars)
}
\arguments{
\item{type}{string of either "hh" or "p" designating to either summarize household populations or individual populations.}

\item{summary_vars}{which variables should we summarize.  The default value is 'base' which includes total number of records.  For households 'base' includes longitude and latitude and the names.  For people, this includes gender.  Otherwise we summarize the base variables and the the variable names which we should summarize.}
}
\value{
var_names a vector of variable names
}
\description{
Get the proper names to subset
}
