% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{get_weight_dists}
\alias{get_weight_dists}
\title{Weight place assignment probabilities}
\usage{
get_weight_dists(dist_mat, places, method = "uniform",
  cap_fxn = cap_default)
}
\arguments{
\item{dist_mat}{a m x n matrix where m is the number of people and 
n is the number of schools}

\item{places}{data frame of places with an ID column}

\item{method}{("uniform", "capacity")}
}
\value{
m x n matrix of probabilities.  Each row should sum to 1
}
\description{
Weight place assignment probabilities
}

