% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getBBox}
\alias{getBBox}
\title{Get the bounding box of a pop}
\usage{
getBBox(hh_pop)
}
\arguments{
\item{hh_pop}{SPEW household population output has longitude and lat parameters}
}
\value{
list with long_min, long_max, lat_min, lat_max
}
\description{
Get the bounding box of a pop
}

