% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getCenters}
\alias{getCenters}
\title{Get the geographic center of a population}
\usage{
getCenters(hh_pop)
}
\arguments{
\item{hh_pop}{SPEW output for household population}
}
\value{
data frame with region name, long and lat
}
\description{
Get the geographic center of a population
}
