% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{getCentersDiags}
\alias{getCentersDiags}
\title{Get the centers df for each region}
\usage{
getCentersDiags(ipums_sum_list)
}
\value{
data frame for plotting
}
\description{
@param ipums_sum_list  output from summarize_ipums()
}
