% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{getLevels}
\alias{getLevels}
\title{Get the number of levels in the structure}
\usage{
getLevels(paths_df)
}
\arguments{
\item{paths_df}{paths df}
}
\value{
number of levels in file path
}
\description{
Get the number of levels in the structure
}
