% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{getVars_ipums}
\alias{getVars_ipums}
\title{Add the base variables to the vars to summarize}
\usage{
getVars_ipums(summary_vars, type)
}
\arguments{
\item{summary_vars}{character vector with variables matching the microdata}

\item{type}{"hh" for household or "p" for people}
}
\value{
list of the categorical and continuous variables na mes
}
\description{
Add the base variables to the vars to summarize
}

