% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\name{haversine}
\alias{haversine}
\title{Get the haversine distance between two points (x1, y1) and (x2, y2) scaled between 0 and 1.}
\usage{
haversine(x1, y1, x2, y2)
}
\arguments{
\item{x1}{longitude of object 1 (vector)}

\item{y1}{latitude of object 1 (vector)}

\item{x2}{longitude of object 2 (vector)}

\item{y2}{latitude of object 2 (vector)}
}
\value{
numeric
}
\description{
Get the haversine distance between two points (x1, y1) and (x2, y2) scaled between 0 and 1.
}
\references{
http://andrew.hedges.name/experiments/haversine/
}

