% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{impute_missing_vals}
\alias{impute_missing_vals}
\title{Impute Missing Values in a data frame}
\usage{
impute_missing_vals(df, method = "mean")
}
\arguments{
\item{df}{where ever column may be imputed (so each column is numeric)}

\item{method}{of imputation, "mean" imputes the mean of the other values into the NA values.  "bootstrap" resamples from non-NA vals.}
}
\value{
df of same dimension, now with no missing vals
}
\description{
Impute Missing Values in a data frame
}
