% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-docs.R
\docType{data}
\name{iso3}
\alias{iso3}
\title{Table of UN iso3 codes}
\format{An object of class \code{data.frame} with 249 rows and 8 columns.}
\usage{
iso3
}
\description{
Table of UN iso3 codes
}
\keyword{datasets}
