% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_characteristics.R
\name{join_pums_to_syneco}
\alias{join_pums_to_syneco}
\title{Connect the synthetic ecosystems to the microdata}
\usage{
join_pums_to_syneco(syneco, pums, join_var, vars_to_join)
}
\arguments{
\item{syneco}{data frame of current syneco}

\item{pums}{the microdata/PUMS we're adding characteristics from}

\item{join_var}{either a string of length 1 with the name of the var to match or a string of length 2 with the first string corresponding to the ID in the syneco and the second from the PUMS}

\item{vars_to_join}{character string of column names of the PUMS to add to the syneco}
}
\value{
A data frame of the syneco with the added characteristics from the microdata
}
\description{
Connect the synthetic ecosystems to the microdata
}
