% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_data}
\alias{make_data}
\title{Create microdata using formatted data}
\usage{
make_data(pop_table, shapefile, pums_h, pums_p, parallel = FALSE,
  sampling_type = "uniform")
}
\arguments{
\item{pop_table}{dataframe with columns corresponding to 
which places need populations, and how many samples to take}

\item{sp}{class object shapefile used for assigning households to 
particular locations}

\item{dataframe}{with microdata corresponding to housegolds}

\item{dataframe}{with microdata corresponding to people}

\item{logical}{indicating whether or not we will generate our 
synthetic populations in parallel}

\item{character}{vector indicating the type of sample to use for 
generating microdata}
}
\value{
logical specifying whether the microdata was generated 
successfully
}
\description{
Create microdata using formatted data
}
\examples{
 make_data(sd_data$pop_table, sd_data$shapefiles, sd_data$pums$pums_h, sd_data$pums$pums_p)
}

