% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_characteristics.R
\name{make_ipf_obj}
\alias{make_ipf_obj}
\title{Set up for creating a set of marginal information for IPF sampling}
\usage{
make_ipf_obj(var_name, type = "ord", bounds, category_name,
  output_file = NULL, df = NULL)
}
\arguments{
\item{var_name}{name of the variable matching the PUMS}

\item{type}{"ord" for ordinal or  "cat" for categorical, the type of variable to be converted to}

\item{bounds}{data frame of upper and lower bounds (inclusive) for variables (numeric)}

\item{output_file}{if not NULL then we save the file as a rds object to output_file}

\item{df}{data frame with place_id and category counts}

\item{category_names}{short name of the category, will be visible to person.  Either length one and the bounds will be pasted to it or length of the number of rows of the bounds with names of your choice.}
}
\value{
a list in the format for use in IPF
}
\description{
Set up for creating a set of marginal information for IPF sampling
}

