% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_maps}
\alias{make_maps}
\title{Make maps for each region/puma ID in a directory}
\usage{
make_maps(output_dir = output_dir, shapefile, pretty = FALSE, zoom = 7,
  parallel = FALSE)
}
\arguments{
\item{output_dir}{path to output directory.  The last node in 
the path will be used as the country name.}

\item{shapefile}{the corresponding country shapefile}

\item{pretty}{if TRUE we try to find the place names instead of the numbers.  Currently nonfunctional}

\item{zoom}{zoom for ggmap 10 is far in, 1 is far out}

\item{parallel}{logical}
}
\value{
logical
}
\description{
Make maps for each region/puma ID in a directory
}
