% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{make_mm_obj}
\alias{make_mm_obj}
\title{Make mm object}
\usage{
make_mm_obj(moments_list, assumption = "independence", nMom = 1,
  type = "cont", region = NULL, path = NULL)
}
\arguments{
\item{moments_list}{a list with each entry as a data frame.  The first df is the first moments, the second the second moments, etc.  Each df has the following format:  place_id | puma_id | var1 moment | var 2 moment |.  The dfs are named mom1, mom2, ...}

\item{assumption}{is either "independent" or "joint"}

\item{nMom}{number of moments}

\item{type}{either "cont" for continuous variable or "ord" for ordered variable}

\item{region}{identifier for region}

\item{path}{if not NULL we will save this object}
}
\value{
list of moment object
}
\description{
Make mm object
}

