% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_place}
\alias{make_place}
\title{Create microdata for individual places}
\usage{
make_place(index, pop_table, shapefile, pums_h, pums_p, sampling_type,
  output_dir)
}
\arguments{
\item{pop_table}{dataframe with columns corresponding to 
which places need populations, and how many samples to take}

\item{shapefile}{sp class object used for assigning households to 
particular locations}

\item{dataframe}{with microdata corresponding to housegolds}

\item{dataframe}{with microdata corresponding to people}

\item{logical}{indicating whether or not we will generate our 
synthetic populations in parallel}

\item{character}{vector indicating the type of sample to use for 
generating microdata}
}
\value{
synthetic population .csv file for both household and person 
level data
}
\description{
Create microdata for individual places
}

