% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-tables.R
\name{makeACSdf}
\alias{makeACSdf}
\title{make a more usable df from acs objects}
\usage{
makeACSdf(acs_obj)
}
\arguments{
\item{acs_obj}{class "acs" from acs package}
}
\value{
a formatted df where the rows are the geographies and the columns are summary table values
}
\description{
make a more usable df from acs objects
}
\details{
to come
}

