% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{makeEnvironmentsDFs}
\alias{makeEnvironmentsDFs}
\title{Test schools/workplaces}
\usage{
makeEnvironmentsDFs(syneco_folder, input_folder, admin_level = 0,
  syneco_vars = c("school_id", "workplace_id", "SCH", "SCHG", "AGEP", "ESR"),
  env_type = "both")
}
\arguments{
\item{syneco_folder}{string}

\item{input_folder}{string}

\item{admin_level}{integer TBD}

\item{syneco_vars}{variables to keep}

\item{env_type}{"sch", "wpl", "both" currently}
}
\value{
list of environmental data frames.  One from the syneco (people), others with input data from schools, workplaces or other
}
\description{
Test schools/workplaces
}

