% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{makeFactors}
\alias{makeFactors}
\title{Turn variables into factors, matching level of PUMS}
\usage{
makeFactors(in_df, my_levels, variables)
}
\arguments{
\item{in_df}{df}

\item{my_levels}{what levels to use}

\item{variables}{the variables to change}
}
\value{
out_df of factors
}
\description{
Turn variables into factors, matching level of PUMS
}

