% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{makeFactorsMarg}
\alias{makeFactorsMarg}
\title{Turn variables into factors, matching level of PUMS}
\usage{
makeFactorsMarg(in_df, marginals, variables)
}
\arguments{
\item{in_df}{df}

\item{variables}{the variables to change}

\item{my_levels}{what levels to use}
}
\value{
out_df of factors
}
\description{
Turn variables into factors, matching level of PUMS
}

