% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-tables.R
\name{makeGeoID}
\alias{makeGeoID}
\title{Make a 11 digit tract ID}
\usage{
makeGeoID(acs_geo)
}
\arguments{
\item{acs_geo}{df from acs object's geography slot}
}
\value{
11 digit ID for each tract
}
\description{
Make a 11 digit tract ID
}
\details{
The geographies should be US tracts
}

