% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{makePlotDF}
\alias{makePlotDF}
\title{Convert ipums_sum_list into a data frame for plotting}
\usage{
makePlotDF(ipums_sum_list)
}
\arguments{
\item{ipums_sum_list}{output from summarize_ipums()}
}
\value{
data frame for plotting
}
\description{
Convert ipums_sum_list into a data frame for plotting
}
