% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{makeStatDF}
\alias{makeStatDF}
\title{Turn the output from test_features into a df}
\usage{
makeStatDF(features_list)
}
\arguments{
\item{test_list}{output from test_features}
}
\value{
a data frame with region id, variables (separated by a dash), chi. sq. value, dof, and p value
}
\description{
Turn the output from test_features into a df
}
