% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{merge_reduce}
\alias{merge_reduce}
\title{Wrapper function for merge}
\usage{
merge_reduce(x, y)
}
\arguments{
\item{x}{x component of merge}

\item{y}{y component of merge}
}
\value{
a function
}
\description{
Wrapper function for merge
}
