% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{partition_pt}
\alias{partition_pt}
\title{Partition the poptable into chunks}
\usage{
partition_pt(total_size, partition_size)
}
\arguments{
\item{partition_size}{indicating the size of chunks 
to partition the pop table into}

\item{total}{size numeric indicating the size of 
the pop table}
}
\description{
Partition the poptable into chunks
}
