% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pathsToDF}
\alias{pathsToDF}
\title{Convert paths to data frame based on folders}
\usage{
pathsToDF(paths)
}
\arguments{
\item{paths}{paths separated by a single /}
}
\value{
df of paths
}
\description{
Convert paths to data frame based on folders
}

