% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.R
\name{people_to_households}
\alias{people_to_households}
\title{Convert a population count to household count}
\usage{
people_to_households(hh_sizes, n_people)
}
\arguments{
\item{hh_sizes}{numeric vector with the household
sizes for this particular sampling region}

\item{n_people}{numeric indicating the number
of people in this specific region}
}
\value{
number of households
}
\description{
Convert a population count to household count
}

