% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{plot_characteristic_proportions}
\alias{plot_characteristic_proportions}
\title{Plot characteristic summary output from summarize_top_spew_region}
\usage{
plot_characteristic_proportions(feature_name = "Feature",
  legend_name = "Types", feature_df, category_names = NULL,
  text_size = 10, region_colors = c("#999999", "#E69F00", "#56B4E9",
  "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7"))
}
\arguments{
\item{feature_name}{string which will be in the title}

\item{legend_name}{name of the legend title}

\item{feature_df}{features summary, output from summarize_top_region()}

\item{category_names}{optional labels to display as the category types.  Default is whatever is contained in the feature_df.}

\item{text_size}{axis text size.  Default is 10}

\item{region_colors}{a string of colors to color the map.  Default is from the colorblind friendly palette.}
}
\value{
a gg map object
}
\description{
Plot characteristic summary output from summarize_top_spew_region
}
