% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_spew.R
\name{plot_labs}
\alias{plot_labs}
\title{Add the labels and the theme to the plot}
\usage{
plot_labs(region_name, g = NULL)
}
\arguments{
\item{region_name}{string, will become the title of the plot
g a ggplot.  Default is NULL.}
}
\description{
Add the labels and the theme to the plot
}
