% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_region}
\alias{plot_region}
\title{Plot the households}
\usage{
plot_region(region_file, addBoundaries = FALSE, map_title = NULL,
  maptype = "roadmap", savePlot = TRUE, output_dir = "../diags/",
  hexBin = FALSE, nBins = 50, alpha = 0.3, subSample = TRUE, K = 10^5,
  zoom = 4)
}
\arguments{
\item{region_file}{file name of SPEW household output}

\item{addBoundaries}{logical indicating whether we should find the corresponding shapefile to add inner region boundaries}

\item{map_title}{default is NULL is also what we save the file as}

\item{maptype}{arguments for ggmap default is "roadmap"}

\item{savePlot}{logical}

\item{nBins}{how many bins should we use}

\item{alpha}{transparency as in ggplot}

\item{subSample}{logical - should we plot the whole thing (possibly very long) or a subset?}

\item{zoom}{as in get_map from ggmap larger number is more zoomed in}

\item{hexbin}{logial - should we use hexagons instead of points?}
}
\value{
a low res png
}
\description{
Plot the households
}

