% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_region_diags}
\alias{plot_region_diags}
\title{Plot the region diagnostics}
\usage{
plot_region_diags(ipums_sum_list, ipums_fs, pretty = TRUE, borders = FALSE,
  data_path = input_dir, savePlot = FALSE, plot_name,
  map_type = "toner-lite", diags_path = ".", type = "ipums", ...)
}
\arguments{
\item{ipums_sum_list}{output from summarize_ipums()}

\item{ipums_fs}{output from summarizeFileStructure()}

\item{pretty}{ggplot or base R}

\item{borders}{logical inclusion of shapefile boundaries}

\item{data_path}{input directory to plot the borders}

\item{savePlot}{logical}

\item{plot_name}{what to save the plot as}

\item{map_type}{for ggplot use, default is toner-lite}

\item{diags_path}{path to diags folder}

\item{type}{either "ipums" or "us"}

\item{...}{plotting params}
}
\value{
logical or ggplot object
}
\description{
Plot the region diagnostics
}
