% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew_in_console.R
\name{plot_syneco}
\alias{plot_syneco}
\title{Plotting the synthetic ecosystem}
\usage{
plot_syneco(input_data, syneco, region_name = NULL, pretty = FALSE, ...)
}
\arguments{
\item{input_data}{a list of the essential data and possibly supplementary data, shapefile must be one of the names}

\item{syneco}{the outputted synthetic ecosystem data from spewR}

\item{region_name}{string}

\item{pretty}{logical if TRUE, we make a ggplot and return it}
}
\value{
plot of the synthetic ecosystem
}
\description{
Plotting the synthetic ecosystem
}

