% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{pretty_print}
\alias{pretty_print}
\title{Print text if doPrint}
\usage{
pretty_print(doPrint = TRUE, text)
}
\arguments{
\item{doPrint}{logical}

\item{text}{string}
}
\value{
print statement
}
\description{
Print text if doPrint
}
