% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{print_region_list}
\alias{print_region_list}
\title{Write out information on each region}
\usage{
print_region_list(region_list)
}
\description{
@param region_list a list containing all of the 
 summary and diagnostic information for each
}

