% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{read_concat_pop}
\alias{read_concat_pop}
\title{Reads in and concatenates a list of filenames and pastes on the header file}
\usage{
read_concat_pop(filenames)
}
\arguments{
\item{filenames}{full paths to files to be concatenated}
}
\value{
dataframe of population
}
\description{
Reads in and concatenates a list of filenames and pastes on the header file
}
