% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data}
\alias{read_data}
\title{Read SPEW input data from files}
\usage{
read_data(base_dir, folders = list(pop_table = NULL, pums = NULL, shapefiles =
  NULL, roads = NULL, schools = NULL, lookup = NULL, workplaces = NULL),
  data_group = "US", vars = list(household = NA, person = NA),
  verbose = FALSE)
}
\arguments{
\item{base_dir}{character vector specifying the ecosystem directory}

\item{folders}{list containing the subdirectory file-path containing where each 
type of data is located}

\item{data_group}{character either "US", "ipums" or "none": Gives the format 
that read_data expects.}

\item{vars}{list with household and person variables specifying which PUMS 
variables to use}
}
\value{
list: Each elements contains a standardized data-source
}
\description{
Based on directory structure for input data on the Olympus computing cluster
}
