% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_moments}
\alias{read_moments}
\title{Read in the R data object for moment matching}
\usage{
read_moments(input_dir, folders, data_group)
}
\arguments{
\item{input_dir}{character vector specifying the directory containing 
all of the input data}

\item{folders}{list which contains the path of each sub-directory with the 
specific data}

\item{data_group}{character either "US", "ipums" or "none" which tells 
read_data if the input data follows a particular format. Used mainly for 
the pre-formatted data-types we have on our Olympus}
}
\value{
moment object
}
\description{
Read in the R data object for moment matching
}
