% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{read_pop}
\alias{read_pop}
\title{Read in the population and subset it to appropriate variables}
\usage{
read_pop(type, input_dir, summary_vars)
}
\arguments{
\item{type}{string of either "hh" or "p" designating to either summarize household populations or individual populations.}

\item{input_dir}{Path to synthetic population outputs generated by generate_spew.  The basename of this dir will be used as the region name unless otherwise specified.}

\item{summary_vars}{which variables should we summarize.  The default value is 'base' which includes total number of records.  For households 'base' includes longitude and latitude and the names.  For people, this includes gender.  Otherwise we summarize the base variables and the the variable names which we should summarize.}

\item{output_dir}{Directory where files should be output.}
}
\value{
pop_list list of  dataframe of the population subsetted to the summary variables and the actual names
}
\description{
Read in the population and subset it to appropriate variables
}

