% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_shapespatial_to_ogr}
\alias{read_shapespatial_to_ogr}
\title{Read in shapefile using readOGR}
\usage{
read_shapespatial_to_ogr(full_path)
}
\arguments{
\item{full_path}{entire path to shapefile ending in .shp, i.e. what you would pass to readShapeSpatial}
}
\value{
the shapefile, class SpatialPolygonsDataFrame
}
\description{
Read in shapefile using readOGR
}
\note{
This is to keep up-to-date and use rgdal:readOGR instead of the now unsupported maptools::readShapeSpatial
}
