% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{readEnvUS}
\alias{readEnvUS}
\title{Read input data for environments for US}
\usage{
readEnvUS(input_folder, type = "sch")
}
\arguments{
\item{input_folder}{filepath}

\item{type}{either "sch" or "wpl" currently}
}
\value{
data frame of schools or workplaces
}
\description{
Read input data for environments for US
}
