% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{readMarginals}
\alias{readMarginals}
\title{Read in marginal objects from the ACS SF}
\usage{
readMarginals(cur_co, marginal_folder, householder_vars)
}
\arguments{
\item{cur_co}{current county character (3 digits)}

\item{marginal_folder}{path to marginals}

\item{householder_vars}{currently in c("NP", "HINCP", "RAC1P", "AGEP")}
}
\value{
list with each entry as list with data frame for the householder_var along with lookup and type
}
\description{
Read in marginal objects from the ACS SF
}
