% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{readSynecos}
\alias{readSynecos}
\title{Read and subset synecos}
\usage{
readSynecos(syneco_folder, type = "people", syneco_vars = c("school_id",
  "workplace_id", "SCH", "SCHG", "AGEP", "ESR"))
}
\arguments{
\item{syneco_folder}{string}

\item{type}{"people" or "household"}

\item{syneco_vars}{vars to subset along with 'place_id'}
}
\value{
list or data frame of input data
}
\description{
Read and subset synecos
}
