% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{remove_count}
\alias{remove_count}
\title{Remove a row from the pop_table}
\usage{
remove_count(pop_table, place)
}
\arguments{
\item{pop_table}{dataframe to update}

\item{place}{character of place_ids to remove}
}
\value{
pop_table updated dataframe with the 
desired place removed
}
\description{
Remove a row from the pop_table
}

