% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-people.R
\name{remove_excess}
\alias{remove_excess}
\title{Remove comma's, accents, etc. from name}
\usage{
remove_excess(name)
}
\arguments{
\item{name}{character}
}
\value{
name with all of the excess baggage removed
}
\description{
Remove comma's, accents, etc. from name
}
