% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{remove_excess_words}
\alias{remove_excess_words}
\title{Remove extraneous words from place names}
\usage{
remove_excess_words(names)
}
\arguments{
\item{names}{character vector of names}
}
\value{
names a character vector of updated names
}
\description{
Remove extraneous words from place names
}

