% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-locations.R
\name{remove_holes}
\alias{remove_holes}
\title{Remove holes from an object of class Polygon}
\usage{
remove_holes(polygon)
}
\arguments{
\item{polygon}{object of class Polygon or Polygons}
}
\value{
polygon without and Polygons with holes
}
\description{
Remove holes from an object of class Polygon
}
\note{
Borrowed the idea from the wild1 package, which 
I wasn't able to load for R 3.2.2, so I found the source code here:
https://github.com/cran/wild1/blob/master/R/remove.holes.r
}
