% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{remove_words}
\alias{remove_words}
\title{Remove excess words}
\usage{
remove_words(word, names)
}
\arguments{
\item{word}{character of the word you want to replace}

\item{names}{character vector of the words 
we can potentially replace}
}
\value{
names character vector with the excess 
word removed
}
\description{
Remove excess words
}
