% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{removeExtraHeaders}
\alias{removeExtraHeaders}
\title{Remove Extra Headers}
\usage{
removeExtraHeaders(hh_pop)
}
\arguments{
\item{hh_pop}{a household population output from SPEW}
}
\value{
hh_pop minus extra headers
}
\description{
Remove Extra Headers
}

