% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{replace_word}
\alias{replace_word}
\title{Replace an existing word}
\usage{
replace_word(word, replace, names)
}
\arguments{
\item{word}{character of the word you want to replace}

\item{replace}{character of what you want to replace 
the word}

\item{names}{character vector of words which we can
replace}
}
\value{
names character vector of the replaced word
}
\description{
Replace an existing word
}
