% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{run_diags}
\alias{run_diags}
\title{Run basic 'sanity check' diagnostics on a folder of SPEW generated populations.  There should only be one household and one people file in each folder.  If not, look into append_pops().}
\usage{
run_diags(input_dir = "./", output_dir = input_dir, save_plots = TRUE,
  pretty = FALSE)
}
\arguments{
\item{input_dir}{path to the directory (bottom level) of populations to summarize}

\item{output_dir}{path to output directory}

\item{save_plots}{boolean indicating whether the plots should be saved}

\item{pretty}{if set to TRUE, should generate a summary pdf including plots (currently non-functional)}
}
\value{
a print out of the files generated
}
\description{
Run basic 'sanity check' diagnostics on a folder of SPEW generated populations.  There should only be one household and one people file in each folder.  If not, look into append_pops().
}

