% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{sample_ipf}
\alias{sample_ipf}
\title{Sample households PUMS accoording to IPD}
\usage{
sample_ipf(n_house, pums_h, pums_p, marginals, alpha = 0, k = 0.001,
  puma_id = NULL, place_id = NULL)
}
\arguments{
\item{n_house}{number of households to sample}

\item{pums_h}{household pums}

\item{pums_p}{people pums, for appending on certain variables}

\item{marginals}{list with characteristics of marginals}

\item{alpha}{number between 0 and 1, weight of categorical variables}

\item{k}{number between 0 and 1, weight of orginal variables}

\item{puma_id}{id indicating the current puma}

\item{place_id}{id indicating the current region}
}
\description{
Sample households PUMS accoording to IPD
}

