% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-locations.R
\name{sample_locations_roads}
\alias{sample_locations_roads}
\title{Sample coordinates from roads}
\usage{
sample_locations_roads(place_id, n_house, shapefile, noise = 1e-04,
  shapefile_id, do_write = TRUE)
}
\arguments{
\item{place_id}{numeric specifiying the ID of the region we are 
subsampling}

\item{n_house}{numeric indicating the number of households}

\item{shapefile}{sp class with all of the locations for each place id.
In addition, we must have road shapefiles so shapefile is a list with both 
the tracts and the roads, tracts is the first object and roads the second.}

\item{noise}{the standard deviation of how much we jitter the road locations in each direction}

\item{do_write}{logical if TRUE, we read in the roads and subset them.  If FALSE, we use the roads provided in shapefile$roads.  Default is TRUE.}
}
\value{
SpatialPoints object with coordinates for the n households
}
\description{
Sample coordinates from roads
}

