% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-locations.R
\name{sample_locations_uniform}
\alias{sample_locations_uniform}
\title{Sample from a particular polygon shapefile}
\usage{
sample_locations_uniform(place_id, n_house, shapefile)
}
\arguments{
\item{place_id}{numeric specifiying the ID of the region we are 
subsampling}

\item{n_house}{numeric indicating the number of households}

\item{shapefile}{sp class with all of the locations for each place id}
}
\value{
SpatialPoints object with coordinates for the n households
}
\description{
Sample from a particular polygon shapefile
}

