% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.R
\name{sample_people}
\alias{sample_people}
\title{Sample from the individual person PUMS data frame}
\usage{
sample_people(household_pums, pums_p)
}
\arguments{
\item{household_pums}{dataframe with the sampled houehold PUMS}

\item{pums_p}{dataframe of the individual microdata}
}
\value{
data indicating the indices of people to sample
}
\description{
Sample from the individual person PUMS data frame
}

