% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-people.R
\name{sample_uniform}
\alias{sample_uniform}
\title{Sample households uniformly}
\usage{
sample_uniform(n_house, pums_h, puma_id = NULL, place_id = NULL)
}
\arguments{
\item{n_house}{number of households to sample}

\item{pums_h}{the household pums}

\item{puma_id}{if specifying the subset of PUMS to sample s}

\item{place_id}{id of the current region}
}
\description{
Sample households uniformly
}
