% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{sample_with_cont}
\alias{sample_with_cont}
\title{Sample from pums}
\usage{
sample_with_cont(pums, table, alpha, k, marginals)
}
\arguments{
\item{pums}{dataframe with marginal columns}

\item{table}{dataframe with marginal combinations and frequencies}

\item{alpha}{number between 0 and 1, weight of categorical variables}

\item{k}{number between 0 and 1, weight of
orginal variables}

\item{marginals}{list with marginal data}
}
\value{
indices of sampled households
}
\description{
Sample from pums
}
