% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{solve_mm_for_var}
\alias{solve_mm_for_var}
\title{Do the MM solving for an individual variable}
\usage{
solve_mm_for_var(var_ind, place_id, mm_row, pums, assumption, meq = 2)
}
\arguments{
\item{var_ind}{variable index of mm_row, colname should match one in PUMS.}

\item{mm_row}{dataframe of 1 row, includes place_id, puma_id, and averages, with names matching PUMS}

\item{pums}{the subsetted PUMS from which the sample will be drawn}

\item{assumption}{"independence"}

\item{meq}{(number of moments + 1) to match (2 is default and corresponds to matching the average)}
}
\value{
x vector of length of the number of records in the PUMS.  These are probabilities for each of the records in the PUMS.
}
\description{
Do the MM solving for an individual variable
}
