% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{solve_mm_weights}
\alias{solve_mm_weights}
\title{Weight the records of the PUMS so the averages in mm_df will be obtained}
\usage{
solve_mm_weights(place_id, mm_row, pums, assumption = "independence",
  meq = 2)
}
\arguments{
\item{place_id}{unique code identifying the place}

\item{mm_row}{dataframe of 1 row, includes place_id, puma_id, and averages, with names matching PUMS}

\item{pums}{the subsetted PUMS from which the sample will be drawn}

\item{assumption}{"independence"}
}
\value{
x vector of length of the number of  records in the PUMS.  These are probabilities for each of the records.
}
\description{
Weight the records of the PUMS so the averages in mm_df will be obtained
}
\details{
the function solve.QP from the "quadprog" package is used.
}
