% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{spew}
\alias{spew}
\title{Generates synthetic ecosysystems}
\usage{
spew(input_dir, folders, data_group, output_dir, parallel = TRUE,
  sampling_type)
}
\arguments{
\item{input_dir}{character vector specifying the directory containing all of the 
input data}

\item{folders}{list which specifies the sub-directory for each piece of data}

\item{data-group}{character vector, either "US", "ipums", or "none".}
}
\value{
logical indicating whether or not this run of spew ended successfully
}
\description{
Generates synthetic ecosysystems
}
\examples{
spew_dir <- system.file("", package="spew")
data_path <- paste0(spew_dir, "/", "data-raw/46")
spew()
}

