% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{spew_mpi}
\alias{spew_mpi}
\title{Run SPEW in Parallel with an MPI backend}
\usage{
spew_mpi(num_places, pop_table, shapefile, pums_h, pums_p, schools, workplaces,
  marginals, sampling_method, locations_method, convert_count, output_dir,
  export_objects, do_subset_pums = TRUE, do_write = TRUE)
}
\description{
Run SPEW in Parallel with an MPI backend
}

