% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{spew_place}
\alias{spew_place}
\title{Generate synthetic ecosystem for single place}
\usage{
spew_place(index, pop_table, shapefile, pums_h, pums_p, schools = NULL,
  workplaces = NULL, marginals = NULL, output_type = "console",
  sampling_method = "uniform", locations_method = "uniform",
  convert_count = FALSE, output_dir = NULL, road_noise = 2e-04,
  timer = TRUE, verbose = TRUE)
}
\arguments{
\item{index}{row of the pop-table to generate}

\item{pop_table}{}

\item{shapefile}{}

\item{pums_h}{}

\item{pums_p}{}

\item{schools}{}

\item{workplaces}{}

\item{marginals}{}

\item{output_type}{}

\item{sampling_method}{}

\item{locations_method}{}

\item{convert_count}{}

\item{output_dir}{}

\item{road_noise}{}

\item{timer}{}

\item{verbose}{}
}
\value{
List of a synthetic ecosystem
}
\description{
Generate synthetic ecosystem for single place
}
