% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{spew_seq}
\alias{spew_seq}
\title{Run SPEW Sequentially}
\usage{
spew_seq(num_places, pop_table, shapefile, pums_h, pums_p, schools, workplaces,
  marginals, output_type, output_dir, convert_count, sampling_method,
  locations_method, road_noise, timer, verbose)
}
\description{
Internal function, only called by the main spew function
}
