% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfiles.R
\name{spewlog_to_df}
\alias{spewlog_to_df}
\title{Convert a SPEW Logfile into a data-frame}
\usage{
spewlog_to_df(logfile, columns, path = NULL)
}
\arguments{
\item{logfile}{character with the file-name}

\item{columns}{character vector indicating the}

\item{path}{path to the logfile (used for debugging) 

names of the columns we are extracting from the
log-file}
}
\description{
Convert a SPEW Logfile into a data-frame
}
