% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew_in_console.R
\name{spewr}
\alias{spewr}
\title{A wrapper function for SPEW to use in console}
\usage{
spewr(pop_table, shapefile, pums_h, pums_p, supplementary_data = NULL,
  sampling_method = "uniform", locations_method = "uniform",
  convert_count = FALSE, do_parallel = FALSE)
}
\arguments{
\item{pop_table}{dataframe with rows corresponding to places which 
need to have a population generated. Columns also required are "puma_id"
and "n_house".}

\item{shapefile}{sp class object used for assigning households to 
particular locations}

\item{pums_h}{dataframe with microdata corresponding to housegolds}

\item{pums_p}{dataframe with microdata corresponding to people}

\item{supplementary_data}{default NULL but otherwise a list of supplementary data which may include schools, workplaces, marginals, or moments.}

\item{sampling_method}{character vector indicating the type of sampling 
method to use, defaults to "uniform". Can also be "ipf" with appropriate marginal data or "mm" with appropriate moments data.}

\item{locations_method}{character vector indicating the type of location 
sampling to use, defaults to "uniform", can also be "roads"}

\item{convert_count}{logical meant to indicate if we are going to convert 
population totals from people to household counts. Default: FALSE, assumes
the population is the total number of households}

\item{do_parallel}{should we run this in parallel with the foreach?  Default is FALSE.  If true, the required packages of 'foreach' and `doParallel` must be installed.}
}
\value{
a synthetic ecosystem with synthetic households, synthetic individuals along with the place_id and puma_id
}
\description{
A wrapper function for SPEW to use in console
}

