% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{spher_dists}
\alias{spher_dists}
\title{Calculates the geodesic distance between two points}
\usage{
spher_dists(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{longitude (degrees)}

\item{lat1}{latitude (degrees)}

\item{long2}{longitude (degrees)}

\item{lat2}{latitude (degrees)}
}
\value{
returns distance in miles
}
\description{
Calculates the geodesic distance between two points
}
\details{
specified by radian latitude/longitude using the
Spherical Law of Cosines (slc)
}

