% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{standardize_pop_table}
\alias{standardize_pop_table}
\title{Make sure pop_table has the appropriate columns}
\usage{
standardize_pop_table(pop_table, data_group)
}
\arguments{
\item{pop_table}{Non-standardized population}

\item{data_group}{character either "US", "ipums" or "none": Gives the format 
that read_data expects.}
}
\description{
Make sure pop_table has the appropriate columns
}
