% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{stat_test_us_joint}
\alias{stat_test_us_joint}
\title{Perform chi square test on tract for joint totals}
\usage{
stat_test_us_joint(regionID, type = "p", level = "tract", output,
  joint_table, variables = c("RAC1P"), marginals)
}
\arguments{
\item{regionID}{11 digit ID for tract, 5 for county, 2 for state.  The FIPS number}

\item{type}{either "p" or "hh" for person or household output, respectively}

\item{output}{SPEW synthetic ecosystem (data frame)}

\item{variables}{variable(s) to test.  No more than pairs}

\item{marginals}{marginal folder to get factor names}

\item{joint_folder}{folder to IPF tables saved from generation}
}
\value{
list of p value for region and original chi square value
}
\description{
Perform chi square test on tract for joint totals
}
\details{
to come
}

