% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{stat_test_us_joint_outer}
\alias{stat_test_us_joint_outer}
\title{Test joint totals to synth pop using chi square}
\usage{
stat_test_us_joint_outer(regionID, type = "b", level = "tract",
  output_folder, joint_folder, variables = c("RAC1P", "AGEP", "HINCP", "NP"),
  puma_id = puma_id, marginals)
}
\arguments{
\item{regionID}{11 digit string}

\item{type}{"b"}

\item{level}{"tract"}

\item{output_folder}{path}

\item{joint_folder}{joint folder to ipf table from syneco generation}

\item{variables}{in c("RAC1P", "AGEP", "HINCP", "NP")}
}
\description{
Test joint totals to synth pop using chi square
}
