% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{stat_test_us_marg}
\alias{stat_test_us_marg}
\title{Perform chi square test on tract for marginals totals}
\usage{
stat_test_us_marg(regionID, type = "p", level = "tract", output, marginals,
  variables = c("RAC1P"))
}
\arguments{
\item{regionID}{11 digit ID for tract, 5 for county, 2 for state.  The FIPS number}

\item{type}{either "p" or "hh" for person or household output, respectively}

\item{output}{SPEW synthetic ecosystem (data frame)}

\item{variables}{variable(s) to test.  No more than pairs}

\item{PUMS}{Public Use Micro Sample to compare joint distributions to}
}
\value{
list of p value for region and original chi square value
}
\description{
Perform chi square test on tract for marginals totals
}
\details{
to come
}
