% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{stat_test_us_marg_outer}
\alias{stat_test_us_marg_outer}
\title{Test Marginal totals to synth pop using chi square}
\usage{
stat_test_us_marg_outer(regionID, type = "b", level = "tract",
  output_folder, marginals, variables = c("RAC1P", "AGEP", "HINCP", "NP"),
  puma_id = puma_id)
}
\arguments{
\item{regionID}{11 digit string}

\item{type}{"b"}

\item{level}{"tract"}

\item{output_folder}{path}

\item{marginals}{output from readMarginals}

\item{variables}{in c("RAC1P", "AGEP", "HINCP", "NP")}
}
\description{
Test Marginal totals to synth pop using chi square
}

