% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{stat_test_us_pums_outer}
\alias{stat_test_us_pums_outer}
\title{Outer function to perform chi square tests}
\usage{
stat_test_us_pums_outer(regionID, type = "p", level = "tract",
  output_folder, PUMS_full, variables = c("RAC1P", "SEX", "AGEP"),
  puma_id = puma_id)
}
\arguments{
\item{regionID}{11 digit ID for tract, 5 for county, 2 for state.  The FIPS number}

\item{type}{either "p" or "hh" for person or household output, respectively}

\item{output_folder}{output_folder (top level)}

\item{PUMS_full}{Public Use Micro Sample (data frame) for whole state}

\item{variables}{variable(s) to test.  Must be for a single test}
}
\value{
list with region ID, variable comparisons, and chi square results
}
\description{
Outer function to perform chi square tests
}
\details{
to come
}
