% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{subset_pums}
\alias{subset_pums}
\title{Align pums with marginal totals}
\usage{
subset_pums(pums_h, pums_p, marginals, puma_id)
}
\arguments{
\item{pums_h}{dataframe of household pums}

\item{pums_p}{dataframe of people pums}

\item{marginals}{list of marginals totals}

\item{puma_id}{id for subsetting the pums}
}
\value{
pums with only relevant
}
\description{
Align pums with marginal totals
}

