% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-locations.R
\name{subset_shapes_roads}
\alias{subset_shapes_roads}
\title{Subset the shapefile and road lines to proper roads within specified tract}
\usage{
subset_shapes_roads(place_id, shapefile)
}
\arguments{
\item{place_id}{numeric specifiying the ID of the region we are 
subsampling}

\item{shapefile}{sp class with all of the locations for each place id.
In addition, we must have road shapefiles so shapefile is a list with both the 
tracts and the roads, tracts is the first object and roads the second.}
}
\value{
new_shp - roads within the tract, a SpatialLines object
}
\description{
Subset the shapefile and road lines to proper roads within specified tract
}

