% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{summarize_environment}
\alias{summarize_environment}
\title{Return the unique environment assignments in a region}
\usage{
summarize_environment(df, env_vars)
}
\arguments{
\item{df}{the dataframe for the region}

\item{env_vars}{the environmental variables we want to summarize}
}
\value{
list of unique environment assignments for each environment
}
\description{
Return the unique environment assignments in a region
}
