% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{summarize_features}
\alias{summarize_features}
\title{Summarize individual features of a region}
\usage{
summarize_features(df, marginals, vars_to_sum)
}
\arguments{
\item{df}{data frame of individuals in region subsetted to vars to sum and marginal equivalents}

\item{marginals}{marginal list or NULL}

\item{vars_to_sum}{name of variables to summarize}
}
\value{
list of table of each feature
}
\description{
Summarize individual features of a region
}
