% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summarize_ipums}
\alias{summarize_ipums}
\title{Summarize ipums populations}
\usage{
summarize_ipums(output_dir, ipums_fs, varsToSummarize = list(vars_hh = "base",
  vars_p = "base"), doPrint = FALSE, sampSize = 10^3, readFun = read.csv)
}
\arguments{
\item{output_dir}{path}

\item{ipums_fs}{output from summarizeFileStructure}

\item{varsToSummarize}{list with first entry as vector of household ipums vars and the second as person ipums vars.  The default value is 'base' which includes total number of records.  For households 'base' includes longitude and latitude and the names.  For people, this includes gender.  Otherwise we summarize the base variables and the the variable names which we should summarize.  For the summary output to be something other than a factor variable then see var_list.}

\item{doPrint}{logical}

\item{sampSize}{number of people to retain (default is 10000) per region for plotting}

\item{readFun}{(how to read in file either read.csv or fread from data.table)}
}
\description{
Summarize ipums populations
}
