% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summarize_pop}
\alias{summarize_pop}
\title{Get the summary values of a spew generated synthetic population and output a row of summary and the column names}
\usage{
summarize_pop(pop_list, region_name)
}
\arguments{
\item{pop_list}{list of  dataframe of the population subsetted to the summary variables and the actual names}

\item{region_name}{string of region_name}
}
\value{
summary_list list of summary numbers with corresponding names
}
\description{
Get the summary values of a spew generated synthetic population and output a row of summary and the column names
}
