% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_spew.R
\name{summarize_spew_region}
\alias{summarize_spew_region}
\title{Summarize a singular region from spew output}
\usage{
summarize_spew_region(spew_region, type = "households", marginals,
  vars_to_sum, vars_to_sum_env = NULL)
}
\arguments{
\item{spew_region}{a single region (entry in the list) from spew}

\item{type}{either "households or "people".  Default is "households".  The type to summarize}

\item{marginals}{marginals object}

\item{vars_to_sum}{names of the variables to summarize}

\item{vars_to_sum}{environment variables to summarize.  Default is NULL}
}
\value{
a list of length of  1 for the  region ID , 1 for population size, and optionally one for a dataframe of stored lat/long coords, the number of variables to summarize , and the environmental variables
}
\description{
Summarize a singular region from spew output
}
