% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew_in_console.R
\name{summarize_syneco}
\alias{summarize_syneco}
\title{Summarize a syneco generated by SPEW}
\usage{
summarize_syneco(syneco, vars_df, make_plots = FALSE, verbose = FALSE)
}
\arguments{
\item{syneco}{the return object from spewr}

\item{vars_df}{a dataframe with 3 columns with names var_name, pop_type, and var_type.  var_name should correspond to a variable within the syneco.  pop_type is either "hh" or "p" indicating whether the variable is to be summarized at the household level or person level.  var_type is either "cont" for continuous variable or "cat" for categorical variable.  This changes how the variable is summarized.}

\item{make_plots}{logical indicating whether we even want plots.}

\item{verbose}{if TRUE then print out summaries.  Default is FALSE}
}
\value{
list of output and plots if pretty = TRUE and make_plots = TRUE
}
\description{
Summarize a syneco generated by SPEW
}

