% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew_in_console.R
\name{summarize_syneco_var}
\alias{summarize_syneco_var}
\title{Summarize a variable}
\usage{
summarize_syneco_var(df, households, people, make_plots = FALSE,
  verbose = FALSE)
}
\arguments{
\item{df}{dataframe with one row with var_name, pop_type, and var_type columns}

\item{households}{synthetic ecosystem of households (dataframe)}

\item{people}{synthetic ecosystem of people (data frame)}

\item{verbose}{if TRUE then print out summaries.  Default is FALSE}

\item{make_plots.}{Default is FALSE.  If TRUE, summarize through plots}
}
\value{
summary of the particular variable
}
\description{
Summarize a variable
}
