% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summarize_us}
\alias{summarize_us}
\title{Summarize us populations}
\usage{
summarize_us(output_dir, us_fs, varsToSummarize = list(vars_hh = "base",
  vars_p = "base"), doPrint = FALSE, sampSize = 10^4, sum_level = 2,
  threshold_list = NULL)
}
\arguments{
\item{output_dir}{path}

\item{us_fs}{output from summarizeFileStructure}

\item{varsToSummarize}{list with first entry as vector of household ipums vars and the second as person ipums vars.  The default value is 'base' which includes total number of records.  For households 'base' includes longitude and latitude and the names.  For people, this includes gender.  Otherwise we summarize the base variables and the the variable names which we should summarize.  For the summary output to be something other than a factor variable then see var_list.}

\item{doPrint}{logical}

\item{sampSize}{number of people to retain (default is 10000) per region for plotting}

\item{sum_level}{1 - state, 2- country or 3- tract - the subregions which we summarize and eventually plot}

\item{threshold_list}{at list of thresholds and names to convert PUMS into factor variables for plotting or use in align_pums}
}
\value{
list
}
\description{
Summarize us populations
}

