% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summarizeFeatures}
\alias{summarizeFeatures}
\title{Summarizing a synthetic ecosystem}
\usage{
summarizeFeatures(var, tab, type = "cat")
}
\arguments{
\item{var}{variable name}

\item{tab}{dataframe of the syneco}

\item{type}{"cat" or "cont" for categorical or continuous data.  right now continuous data is not summarized}
}
\value{
summary info
}
\description{
Summarizing a synthetic ecosystem
}

