% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summarizeFileStructure}
\alias{summarizeFileStructure}
\title{Sumarize the files within the output dir}
\usage{
summarizeFileStructure(output_dir, doPrint = FALSE, type = "ipums")
}
\arguments{
\item{output_dir}{path to output of SPEW region name}

\item{doPrint}{logical}

\item{type}{either "ipums" or "us"}
}
\value{
list
}
\description{
Sumarize the files within the output dir
}

