% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summarizeLowestFS}
\alias{summarizeLowestFS}
\title{Build the FS for a lowest level region}
\usage{
summarizeLowestFS(tract_id, output_dir, doPrint = FALSE, type = "us")
}
\arguments{
\item{tract_id}{11 digits for US (character)}

\item{output_dir}{path to output of SPEW's lowest level (e.g. this folder as csvs from SPEW)}

\item{doPrint}{logical}

\item{type}{either "ipums" or "us"}
}
\value{
list (same format as summarizeFileStructure)
}
\description{
#' Sumarize the files within the output dir
}

