% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-docs.R
\docType{data}
\name{tartanville}
\alias{tartanville}
\title{Input data for Tartanville}
\format{An object of class \code{list} of length 6.}
\usage{
tartanville
}
\description{
Input data for Tartanville
}
\keyword{datasets}
