% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{test_features}
\alias{test_features}
\title{Loop over regions in a state folder to perform chi square tests}
\usage{
test_features(output_folder, PUMS_folder, household_vars = c("HINCP", "NP"),
  people_vars = c("RAC1P", "SEX"), householder_vars = c("RAC1P", "AGEP",
  "HINCP", "NP"))
}
\arguments{
\item{output_folder}{output_folder (top level)}

\item{PUMS_folder}{folder to  Public Use Micro Sample (data frame) for whole state}

\item{household_vars}{string of household variables to test}

\item{people_vars}{string of people vars to test}

\item{householder_vars}{string of head of householder vars to test}
}
\value{
list with region ID, variable comparisons, and chi square results
}
\description{
Loop over regions in a state folder to perform chi square tests
}
\details{
to come
}
