% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{test_features_marg}
\alias{test_features_marg}
\title{Loop over regions in a state folder to perform chi square tests}
\usage{
test_features_marg(output_folder, marginal_folder, householder_vars = NULL)
}
\arguments{
\item{output_folder}{output_folder (top level)}

\item{marginal_folder}{folder to  marginal SF}

\item{householder_vars}{string of head of householder vars to test}

\item{household_vars}{string of household variables to test}

\item{people_vars}{string of people vars to test}
}
\value{
list with region ID, variable comparisons, and chi square results
}
\description{
Loop over regions in a state folder to perform chi square tests
}
\details{
to come
}
