% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{update_freqs}
\alias{update_freqs}
\title{Update frequencies to match # of households}
\usage{
update_freqs(freqs, n_house)
}
\arguments{
\item{n_house}{number of households}

\item{frequencies}{}
}
\description{
Update frequencies to match # of households
}
