% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-docs.R
\docType{data}
\name{us}
\alias{us}
\title{Table of US states and counties}
\format{An object of class \code{data.frame} with 3235 rows and 8 columns.}
\usage{
us
}
\description{
Table of US states and counties
}
\keyword{datasets}
