% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-docs.R
\docType{data}
\name{us_pums_sf}
\alias{us_pums_sf}
\title{An example marginal distribution table}
\format{An object of class \code{list} of length 4.}
\usage{
us_pums_sf
}
\description{
An example marginal distribution table
}
\keyword{datasets}
