% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfiles.R
\name{verify_column}
\alias{verify_column}
\title{Verify the column is the correct size}
\usage{
verify_column(column, df_size)
}
\arguments{
\item{column}{to be verified}

\item{df_size}{size}
}
\description{
Verify the column is the correct size
}
