% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\name{weight_dists2}
\alias{weight_dists2}
\title{Weight school assignment probabilities}
\usage{
weight_dists2(dist_mat, schools)
}
\arguments{
\item{dist_mat}{a m x n matrix where m is the number of people and 
n is the number of schools}

\item{schools}{data frame of schools}
}
\value{
m x n matrix of probabilities.  Each row should sum to 1
}
\description{
Weight school assignment probabilities
}
