% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\name{weight_dists_D}
\alias{weight_dists_D}
\title{Weight school assignment probabilities, distance only}
\usage{
weight_dists_D(dist_mat, schools)
}
\arguments{
\item{dist_mat}{a m x n matrix where m is the number of people and 
n is the number of schools}

\item{schools_sub}{data frame of schools}
}
\value{
m x n matrix of probabilities.  Each row should sum to 1
}
\description{
Weight school assignment probabilities, distance only
}

