% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{write_pop_table}
\alias{write_pop_table}
\title{Write out the final, formatted table}
\usage{
write_pop_table(pop_table, output_dir)
}
\arguments{
\item{pop_table}{the population table df}

\item{output_dir}{character vector of the directory 
to write the population table}
}
\value{
logical TRUE if completed. As well as a written 
pop_table to the given output directory
}
\description{
Write out the final, formatted table
}

