% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{write_schools}
\alias{write_schools}
\title{Write school environment}
\usage{
write_schools(schools, env_dir)
}
\arguments{
\item{schools}{school data-object}

\item{env_dir}{filepath to write out the environment}
}
\value{
logical TRUE if completed successfully
}
\description{
Write school environment
}

