% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{write_workplaces}
\alias{write_workplaces}
\title{Write workplaces environment}
\usage{
write_workplaces(workplaces, env_dir)
}
\arguments{
\item{workplaces}{school data-object}

\item{env_dir}{filepath to write out the environment}
}
\value{
logical TRUE if completed successfully
}
\description{
Write workplaces environment
}

