% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{assessSchUS}
\alias{assessSchUS}
\title{Assess the schools of the us against the agents}
\usage{
assessSchUS(agents, schools_pub, schools_priv, distFun = haversine)
}
\description{
Assess the schools of the us against the agents
}
