% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{calc_dists}
\alias{calc_dists}
\title{Calculate distance b/w cont table row and pums}
\usage{
calc_dists(pums, table_row, alpha, k, marginals)
}
\arguments{
\item{table_row}{}
}
\description{
Calculate distance b/w cont table row and pums
}

