% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{call_spew}
\alias{call_spew}
\title{Wrapper function to read, format, and generate SPEW synthetic ecosystems}
\usage{
call_spew(base_dir, folders = NULL, data_group = "US",
  parallel_type = "SEQ", sampling_method = "uniform",
  locations_method = "uniform", convert_count = FALSE,
  vars = list(household = NA, person = NA), do_subset_pums = TRUE,
  do_write = TRUE, noise = 2e-04)
}
\arguments{
\item{base_dir}{character specifying ecosystem directory}

\item{folders}{list specifying sub-directories for each data-source}

\item{parallel_type}{Indicating to run sequentially or parallel. If parallel, 
the back-end is specified, either "MPI", "SOCK", or "MC"}

\item{sampling_method}{character of sampling method. Default: "uniform", can also 
be "ipf" or "mm" if appropriate marginal/moments data is included}

\item{locations_method}{character vector indicating the type of location 
sampling to use. Default to "uniform". can also be "roads" if appropriate line 
shapefiles are included.}

\item{convert_count}{logical meant to indicate if we are going to convert 
population totals from people to household counts. Default: FALSE, assumes
the population is the total number of households}

\item{vars}{list with two elements: household and person. This specifies 
which variables to include in the corresponding household and person PUMS data-set}

\item{do_subset_pums}{logical, do we need to subset PUMS when doing the method of sampling.  Default is TRUE.}

\item{do_write}{logical, do we need to write out the synthetic populations and environments?
If FALSE, then we instead return the synthetic households and people in a list.  Default is TRUE.}

\item{noise}{a numeric value indicating the value of the noise we add to household locations during road-based sampling.  Default is .0002.}

\item{data-group}{character, either "US", "ipums", or "none"}
}
\value{
logical indicating whether or not this run of spew ended successfully
}
\description{
Wrapper function to read, format, and generate SPEW synthetic ecosystems
}

