% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.R
\name{make_data}
\alias{make_data}
\title{Create synthetic microdata}
\usage{
make_data(pop_table, shapefile, pums_h, pums_p, parallel = FALSE,
  sampling_type = "uniform", output_dir = "/home/lee/south_dakota/",
  convert_count, make_plots = FALSE)
}
\arguments{
\item{pop_table}{dataframe with columns corresponding to
which places need populations, and how many samples to take}

\item{shapefile}{sp class object used for assigning households to
particular locations}

\item{pums_h}{dataframe with microdata corresponding to housegolds}

\item{pums_p}{dataframe with microdata corresponding to people}

\item{parallel}{logical indicating whether or not we will generate our
synthetic populations in parallel}

\item{sampling_type}{character vector indicating the type oof sampling used.
Default's to "uniform"}

\item{output_dir}{character vector specifying where to write the synthetic microdata}
}
\value{
logical specifying whether the microdata was generated
successfully
}
\description{
Create synthetic microdata
}
\examples{
make_data(sd_data$pop_table, sd_data$shapefiles, sd_data$pums$pums_h, sd_data$pums$pums_p)
}

