% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_data}
\alias{make_data}
\title{Create synthetic microdata}
\usage{
make_data(pop_table, shapefile, pums_h, pums_p, schools, workplaces, marginals,
  convert_count, output_dir, parallel = FALSE, sampling_method = "uniform",
  locations_method = "uniform", outfile_loc = "")
}
\arguments{
\item{pop_table}{dataframe with columns corresponding to 
which places need populations, and how many samples to take}

\item{shapefile}{sp class object used for assigning households to 
particular locations}

\item{pums_h}{dataframe with microdata corresponding to housegolds}

\item{pums_p}{dataframe with microdata corresponding to people}

\item{schools}{list with names "public" and "private" with a 
dataframe of schools corresponding to public or private, respectively}

\item{workplaces}{dataframe of workplaces with a workplace_id column, 
employees column, and stcotr column}

\item{marginals}{list of marginal population totals. Each element of the 
list contains the marginal totals of a separate variable}

\item{output_dir}{character vector specifying where to write the synthetic microdata}

\item{parallel}{logical indicating whether or not we will generate our 
synthetic populations in parallel}

\item{sampling_method}{character vector indicating the type of sampling 
method to use, defaults to "uniform"}

\item{locations_method}{character vector indicating the type of location 
sampling to use, defaults to "uniform", can also be "roads"}
}
\value{
logical specifying whether the microdata was generated 
successfully
}
\description{
Create synthetic microdata
}
\examples{
make_data(sd_data$pop_table, sd_data$shapefiles, sd_data$pums$pums_h, sd_data$pums$pums_p)
}

