% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_place}
\alias{make_place}
\title{Create microdata for individual places}
\usage{
make_place(index, pop_table, shapefile, pums_h, pums_p, schools, workplaces,
  output_dir, convert_count, sampling_method, locations_method)
}
\arguments{
\item{index}{numeric with the pop_table row to generate}

\item{pop_table}{dataframe with columns corresponding to 
which places need populations, and how many samples to take}

\item{shapefile}{sp class object used for assigning households to 
particular locations}

\item{pums_h}{dataframe with microdata corresponding to housegolds}

\item{pums_p}{dataframe with microdata corresponding to people}

\item{schools}{dataframe with data corresponding to available schools}

\item{workplaces}{dataframe with data corresponding to available workplaces}

\item{sampling_method}{character vector indicating the type of sampling 
method to use, defaults to "uniform"}

\item{locations_method}{character vector indicating the type of location 
sampling to use, defaults to "uniform", can also be "roads"}

\item{output}{dir character vector containing the location to save the}
}
\value{
synthetic population .csv file for both household and person 
level data
}
\description{
Create microdata for individual places
}

