% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_place}
\alias{make_place}
\title{Create microdata for individual places}
\usage{
make_place(index, pop_table, shapefile, pums_h, pums_p, sampling_type,
  output_dir, convert_count, make_plots = FALSE)
}
\arguments{
\item{index}{numeric with the pop_table row to generate}

\item{pop_table}{dataframe with columns corresponding to 
which places need populations, and how many samples to take}

\item{shapefile}{sp class object used for assigning households to 
particular locations}

\item{pums_h}{dataframe with microdata corresponding to housegolds}

\item{pums_p}{dataframe with microdata corresponding to people}

\item{sampling_type}{character vector indicating the type of sample to use for 
generating microdata. Right now the only value here is "uniform"}

\item{make_plots}{boolean indicating whether spew  makes plots of the synthetic households
synthetic population}

\item{output}{dir character vector containing the location to save the}
}
\value{
synthetic population .csv file for both household and person 
level data
}
\description{
Create microdata for individual places
}

