% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pop}
\alias{plot_pop}
\title{Plot the synthetic household populations}
\usage{
plot_pop(region_name, hh_pop, region_shape, zoom = 8,
  title_map = region_name)
}
\arguments{
\item{region_name}{character name of the region for use in 
finding with ggmap}

\item{hh_pop}{the synthetic population generated by spew for a 
given region}

\item{region_shape}{shapefile of the region ONLY}

\item{zoom}{zoom for ggmap 10 is far in, 1 is far out}

\item{title_map}{- title for the map}
}
\value{
plot of the region, a ggmap object
}
\description{
Plot the synthetic household populations
}

