% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\name{plot_region}
\alias{plot_region}
\title{Plot SPEW region}
\usage{
plot_region(coords_df, get_world_map = FALSE, f = 0.1,
  region_colors = c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442",
  "#0072B2", "#D55E00", "#CC79A7"))
}
\arguments{
\item{coords_df}{data frame with "longitude", "latitude", and "region_id" columns}

\item{get_world_map}{logical.  Get an underlying map from ggmap.  Default is TRUE}

\item{f}{border size between 0 and 1.  Default is .1.}

\item{region_colors}{a string of colors to color the map.  Default is from the colorblind friendly palette.}
}
\value{
a ggmap object
}
\description{
Plot SPEW region
}
