% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data}
\alias{read_data}
\title{Read in the SPEW input data}
\usage{
read_data(base_dir, folders = list(pop_table = NULL, pums = NULL, shapefiles =
  NULL, roads = NULL, schools = NULL, lookup = NULL, workplaces = NULL),
  data_group = "US", vars = list(household = NA, person = NA))
}
\arguments{
\item{base_dir}{character vector specifying the ecosystem directory}

\item{folders}{list containing the path of each data sub directory}

\item{data_group}{character either "US", "ipums" or "none" which tells 
read_data if the input data follows a particular format. Used mainly for 
the pre-formatted data-types we have on our Olympus}

\item{vars}{list with two components: household and person. This specifies 
which variables to include in the corresponding PUMS data-set}
}
\value{
list in which each element contains one of our standardized 
data sources
}
\description{
Based on the hierarchy on the Olympus Supercomputer
}

