% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data}
\alias{read_data}
\title{Read in all necessary data}
\usage{
read_data(input_dir, folders = list(pop_table = "popTables", pums =
  "pums/2013", schools = "schools", lookup = "tables", shapefiles = "tiger",
  workplaces = "workplaces"), data_group = "US", vars = list(household = NA,
  person = NA))
}
\arguments{
\item{input_dir}{character vector specifying the directory containing 
all of the input data}

\item{folders}{list which contains the path of each sub-directory with the 
specific data}

\item{data_group}{character either "US", "ipums" or "none" which tells 
read_data if the input data follows a particular format. Used mainly for 
the pre-formatted data-types we have on our Olympus}

\item{vars}{list with two components: household and person. This specifies 
which variables to include in the corresponding PUMS data-set}
}
\value{
list in which each element contains one of our standardized 
data sources
}
\description{
Read in all necessary data
}

