% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-locations.R
\name{samp_roads}
\alias{samp_roads}
\title{Sample the locations from the liens of a SpatialLines object}
\usage{
samp_roads(n_house, new_shp, noise)
}
\arguments{
\item{n_house}{number of households}

\item{new_shp}{SpatialLines object}

\item{noise}{std deviation of Gaussian noise added to coordinates, default is .001}
}
\value{
SpatialPoints object with coordinates for the n_house households
}
\description{
Sample the locations from the liens of a SpatialLines object
}

