% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{sample_households}
\alias{sample_households}
\title{Sample appropriate indices from household PUMS}
\usage{
sample_households(n_house, pums_h, puma_id = NULL,
  sampling_type = "uniform")
}
\arguments{
\item{n_house}{numeric indicating the number of households to sample}

\item{pums_h}{dataframe of the households we are sampling from}

\item{puma_id}{vector indicating which specific puma in PUMS we are sampling 
from, if any}

\item{sampling_type}{character indicating the method for sampling}
}
\value{
numeric with the indicies of the household PUMS to sample
}
\description{
Sample appropriate indices from household PUMS
}

