% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{sample_mm}
\alias{sample_mm}
\title{Sample households PUMS according to MM}
\usage{
sample_mm(n_house, pums_h, pums_p, mm_obj, puma_id = NULL, place_id = NULL)
}
\arguments{
\item{n_house}{number of households to sample}

\item{pums_h}{household pums}

\item{pums_p}{people pums, for appending on certain variables}

\item{mm_obj}{list with moment matching data}

\item{puma_id}{id indicating the current puma}

\item{place_id}{id indicating the current region}
}
\description{
Sample households PUMS according to MM
}
