% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-people.R
\name{sample_people}
\alias{sample_people}
\title{Sample from the individual person PUMS data frame}
\usage{
sample_people(method, household_pums, pums_p)
}
\arguments{
\item{method}{character indicating the method for sampling}

\item{household_pums}{dataframe with the sampled houehold PUMS}

\item{pums_p}{dataframe of the individual microdata}
}
\value{
people numeric vector indicating the indices of people to sample
}
\description{
Sample from the individual person PUMS data frame
}

