\name{spew-package}
\alias{spew-package}
\alias{spew}
\docType{package}
\title{
\packageTitle{spew}
}
\description{
\packageDescription{spew}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spew}
\packageIndices{spew}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{spew}

Maintainer: \packageMaintainer{spew}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
