% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{spew}
\alias{spew}
\alias{spew-alg}
\title{SPEW algorithm to generate synthetic ecosystems}
\usage{
spew(pop_table, shapefile, pums_h, pums_p, schools = NULL,
  workplaces = NULL, marginals = NULL, output_type = "console",
  base_dir = NULL, convert_count = FALSE, run_type = "SEQ",
  sampling_method = "uniform", locations_method = "uniform",
  outfile_loc = "", road_noise = 2e-04, timer = TRUE, verbose = TRUE)
}
\arguments{
\item{pop_table}{dataframe where rows correspond to places where populations 
should be generated. Other requird columns are "n_house" and "puma_id"}

\item{shapefile}{sp class object used for assigning households to 
particular locations}

\item{pums_h}{dataframe with microdata corresponding to housegolds}

\item{pums_p}{dataframe with microdata corresponding to people}

\item{schools}{list with names "public" and "private" with a 
dataframe of schools corresponding to public or private, respectively}

\item{workplaces}{dataframe of workplaces with a workplace_id column, 
employees column, and stcotr column}

\item{marginals}{list of marginal population totals. Each element of the 
list contains the marginal totals of a separate variable}

\item{output_type}{Default is "console" if we want to resulting population 
as an R variable. Alternative is "write", which is used on Olympus for writing 
out .csv files of the population}

\item{base_dir}{character specifying ecosystem directory write to. Only used if
output_type = "write"}

\item{convert_count}{logical meant to indicate if we are going to convert 
population totals from people to household counts. Default: FALSE, assumes
the population is the total number of households}

\item{run_type}{Whether to run sequentially in parallel. Default is "SEQ", for 
a sequential run. If parallel, back-end is either "MPI", "SOCK", or "MC"}

\item{sampling_method}{character indicating the type of sampling 
method to use, defaults to "uniform". Can also be "ipf" with appropriate marginal data.}

\item{locations_method}{character indicating the type of location 
sampling to use, defaults to "uniform", can also be "roads".}

\item{outfile_loc}{Defaults to "", so we print out the parallel run information. 
Only set to "/dev/null" for internal testing putposes.}

\item{road_noise}{Noise added to households during road-based sampling}

\item{timer}{logical indicating we want to time the run}

\item{verbose}{logical indicating we want to print output during the run}
}
\value{
logical indicating whether or not this run of spew ended successfully
}
\description{
SPEW algorithm to generate synthetic ecosystems
}
