% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{spew}
\alias{spew}
\title{SPEW algorithm to generate synthetic ecosystems}
\usage{
spew(base_dir, pop_table, shapefile, pums_h, pums_p, schools, workplaces,
  marginals, convert_count, parallel_type = "SEQ",
  sampling_method = "uniform", locations_method = "uniform",
  outfile_loc = "", do_subset_pums = TRUE, do_write = TRUE,
  noise = 2e-04)
}
\arguments{
\item{base_dir}{character specifying ecosystem directory}

\item{pop_table}{dataframe with rows corresponding to places which 
need to have a population generated. Columns also required are "puma_id"
and "n_house".}

\item{shapefile}{sp class object used for assigning households to 
particular locations}

\item{pums_h}{dataframe with microdata corresponding to housegolds}

\item{pums_p}{dataframe with microdata corresponding to people}

\item{schools}{list with names "public" and "private" with a 
dataframe of schools corresponding to public or private, respectively}

\item{workplaces}{dataframe of workplaces with a workplace_id column, 
employees column, and stcotr column}

\item{marginals}{list of marginal population totals. Each element of the 
list contains the marginal totals of a separate variable}

\item{convert_count}{logical meant to indicate if we are going to convert 
population totals from people to household counts. Default: FALSE, assumes
the population is the total number of households}

\item{parallel_type}{Indicating to run sequentially or parallel. Default: "SEQ". 
If parallel, the back-end is specified, either "MPI", "SOCK", or "MC"}

\item{sampling_method}{character vector indicating the type of sampling 
method to use, defaults to "uniform". Can also be "ipf" with appropriate marginal data.}

\item{locations_method}{character vector indicating the type of location 
sampling to use, defaults to "uniform", can also be "roads".}

\item{outfile_loc}{Defaults to "", so we print out the parallel run information. 
Only set to "/dev/null" for internal testing putposes.}

\item{do_subset_pums}{logical, do we need to subset PUMS when doing the method of sampling.  Default is TRUE.}

\item{do_write}{logical, do we need to write out the synthetic populations and environments?
If FALSE, then we instead return the synthetic households and people in a list.  Default is TRUE.}

\item{noise}{a numeric value indicating the value of the noise we add to household locations during road-based sampling.  Default is .0002.}
}
\value{
logical indicating whether or not this run of spew ended successfully
}
\description{
SPEW algorithm to generate synthetic ecosystems
}

