% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{spew_place}
\alias{spew_place}
\title{Create microdata for individual places}
\usage{
spew_place(index, pop_table, shapefile, pums_h, pums_p, schools, workplaces,
  marginals, output_dir, convert_count, sampling_method, locations_method,
  do_subset_pums = TRUE, do_write = TRUE, noise = 2e-04)
}
\arguments{
\item{index}{numeric with the pop_table row to generate}

\item{pop_table}{dataframe with columns corresponding to 
which places need populations, and how many samples to take}

\item{shapefile}{sp class object used for assigning households to 
particular locations}

\item{pums_h}{dataframe with microdata corresponding to housegolds}

\item{pums_p}{dataframe with microdata corresponding to people}

\item{schools}{dataframe with data corresponding to available schools}

\item{workplaces}{dataframe with data corresponding to available workplaces}

\item{marginals}{list with elements corresponding to marginal totals 
of population variables}

\item{output_dir}{character vector containing the location to save the}

\item{sampling_method}{character vector indicating the type of sampling 
method to use, defaults to "uniform".  Other options are "ipf" and "mm"}

\item{locations_method}{character vector indicating the type of location 
sampling to use, defaults to "uniform", can also be "roads"}

\item{do_subset_pums}{logical, do we need to subset PUMS when doing the method of sampling.  Default is TRUE.}

\item{do_write}{logical, do we need to write out the synthetic populations and environments?}

\item{noise}{a numeric value indicating the value of the noise we add to household locations during road-based sampling.  Default is .0002.
If FALSE, then we instead return the synthetic households and people in a list.  Default is TRUE.}
}
\value{
synthetic population .csv file for both household and person 
level data
}
\description{
Create microdata for individual places
}

