% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew.R
\name{spew_seq}
\alias{spew_seq}
\title{Run SPEW Sequentially}
\usage{
spew_seq(num_places, pop_table, shapefile, pums_h, pums_p, schools, workplaces,
  marginals, sampling_method, locations_method, convert_count, output_dir,
  do_subset_pums = TRUE, do_write = TRUE)
}
\description{
Run SPEW Sequentially
}

