% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{standardize_pop_table}
\alias{standardize_pop_table}
\title{Make sure pop_table has the appropriate columns}
\usage{
standardize_pop_table(pop_table, data_group)
}
\description{
Make sure pop_table has the appropriate columns
}

