% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spew_in_console.R, R/summarize_spew.R
\name{summarize_syneco}
\alias{summarize_syneco}
\alias{summarize_syneco}
\title{Summarize a syneco generated by SPEW}
\usage{
summarize_syneco(syneco, vars_to_sum_h, vars_to_sum_p, vars_to_sum_env = NULL,
  samp_size = 10^4, marginals = NULL)

summarize_syneco(syneco, vars_to_sum_h, vars_to_sum_p, vars_to_sum_env = NULL,
  samp_size = 10^4, marginals = NULL)
}
\arguments{
\item{syneco}{the return object from spewr}

\item{vars_to_sum_h}{character vector of variables from the household data frame output to summarize}

\item{vars_to_sum_p}{character vector of variables from the person data frame output to summarize}

\item{vars_to_sum_env}{character vector of variables from the person data frame which correspond to environment assignments.  Default is NULL.}

\item{samp_size}{number of agents to retain from each lower-level region, for plotting purposes only.  Default is 10^4.}

\item{marginals}{list containing all of the marginal totals.  See ?make_ipf_marg for more details.}

\item{vars_df}{a dataframe with 3 columns with names "var_name", "pop_type", and "var_type".  var_name should correspond to a variable within the syneco.  pop_type is either "hh" or "p" indicating whether the variable is to be summarized at the household level or person level.  var_type is either "cont" for continuous variable or "cat" for categorical variable.  This changes how the variable is summarized.}

\item{make_plots}{logical indicating whether we want plots of the individual variables.  Currently is not functional.}

\item{verbose}{if TRUE then print out summaries.  Default is FALSE}

\item{syneco}{output from the 'spew' function}
}
\value{
list of output summaries

list with the household summary list, people summary list, header for households, and header for people, and a data frame of plotting coordinates by summary region
}
\description{
Summarize a syneco generated by SPEW

Summarize synthetic ecosystem for SPEW console output
}
