% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{write_data}
\alias{write_data}
\title{Output our final synthetic populations as csv's}
\usage{
write_data(df, place_id, type, output_dir)
}
\arguments{
\item{df}{dataframe with the final synthetic population}

\item{place_id}{numeric indicating the name of the particular region samples}

\item{type}{character vector with the type, either "household" or "people"}

\item{output_dir}{character containing the directory in which we want to 
write the final csv's}
}
\value{
data indicating the indices of people to sample
}
\description{
Output our final synthetic populations as csv's
}

